% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{accept_vpc_endpoint_connections}
\alias{accept_vpc_endpoint_connections}
\title{Accepts one or more interface VPC endpoint connection requests to your VPC endpoint service}
\usage{
accept_vpc_endpoint_connections(ServiceId, VpcEndpointIds, DryRun = NULL)
}
\arguments{
\item{ServiceId}{The ID of the endpoint service.}

\item{VpcEndpointIds}{The IDs of one or more interface VPC endpoints.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Accepts one or more interface VPC endpoint connection requests to your VPC endpoint service.
}
\examples{

}
