% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{accept_vpc_peering_connection}
\alias{accept_vpc_peering_connection}
\title{Accept a VPC peering connection request}
\usage{
accept_vpc_peering_connection(DryRun = NULL,
  VpcPeeringConnectionId = NULL)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{VpcPeeringConnectionId}{The ID of the VPC peering connection. You must specify this parameter in the request.}
}
\description{
Accept a VPC peering connection request. To accept a request, the VPC peering connection must be in the \code{pending-acceptance} state, and you must be the owner of the peer VPC. Use DescribeVpcPeeringConnections to view your outstanding VPC peering connection requests.
}
\details{
For an inter-region VPC peering connection request, you must accept the VPC peering connection in the region of the accepter VPC.
}
\section{Accepted Parameters}{
\preformatted{accept_vpc_peering_connection(
  DryRun = TRUE|FALSE,
  VpcPeeringConnectionId = "string"
)
}
}

