% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{associate_iam_instance_profile}
\alias{associate_iam_instance_profile}
\title{Associates an IAM instance profile with a running or stopped instance}
\usage{
associate_iam_instance_profile(IamInstanceProfile, InstanceId)
}
\arguments{
\item{IamInstanceProfile}{The IAM instance profile.}

\item{InstanceId}{The ID of the instance.}
}
\description{
Associates an IAM instance profile with a running or stopped instance. You cannot associate more than one IAM instance profile with an instance.
}
\examples{
# This example associates an IAM instance profile named admin-role with
# the specified instance.
associate_iam_instance_profile(
  IamInstanceProfile = list(
    Name = "admin-role"
  ),
  InstanceId = "i-123456789abcde123"
)

}
