% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{attach_network_interface}
\alias{attach_network_interface}
\title{Attaches a network interface to an instance}
\usage{
attach_network_interface(DeviceIndex, InstanceId, NetworkInterfaceId,
  DryRun = NULL)
}
\arguments{
\item{DeviceIndex}{The index of the device for the network interface attachment.}

\item{InstanceId}{The ID of the instance.}

\item{NetworkInterfaceId}{The ID of the network interface.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Attaches a network interface to an instance.
}
\examples{
# This example attaches the specified network interface to the specified
# instance.
attach_network_interface(
  DeviceIndex = 1L,
  InstanceId = "i-1234567890abcdef0",
  NetworkInterfaceId = "eni-e5aa89a3"
)

}
