% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{attach_volume}
\alias{attach_volume}
\title{Attaches an EBS volume to a running or stopped instance and exposes it to the instance with the specified device name}
\usage{
attach_volume(
  Device = "string",
  InstanceId = "string",
  VolumeId = "string",
  DryRun = TRUE|FALSE
)
}
\arguments{
\item{Device}{[required] The device name (for example, \code{/dev/sdh} or \code{xvdh}).}

\item{InstanceId}{[required] The ID of the instance.}

\item{VolumeId}{[required] The ID of the EBS volume. The volume and instance must be within the same Availability Zone.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Attaches an EBS volume to a running or stopped instance and exposes it to the instance with the specified device name.
}
\details{
Encrypted EBS volumes may only be attached to instances that support Amazon EBS encryption. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/EBSEncryption.html}{Amazon EBS Encryption} in the \emph{Amazon Elastic Compute Cloud User Guide}.

For a list of supported device names, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ebs-attaching-volume.html}{Attaching an EBS Volume to an Instance}. Any device names that aren't reserved for instance store volumes can be used for EBS volumes. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/InstanceStorage.html}{Amazon EC2 Instance Store} in the \emph{Amazon Elastic Compute Cloud User Guide}.

If a volume has an AWS Marketplace product code:
\itemize{
\item The volume can be attached only to a stopped instance.
\item AWS Marketplace product codes are copied from the volume to the instance.
\item You must be subscribed to the product.
\item The instance type and operating system of the instance must support the product. For example, you can't detach a volume from a Windows instance and attach it to a Linux instance.
}

For more information about EBS volumes, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ebs-attaching-volume.html}{Attaching Amazon EBS Volumes} in the \emph{Amazon Elastic Compute Cloud User Guide}.
}
\examples{
# This example attaches a volume (``vol-1234567890abcdef0``) to an
# instance (``i-01474ef662b89480``) as ``/dev/sdf``.
attach_volume(
  Device = "/dev/sdf",
  InstanceId = "i-01474ef662b89480",
  VolumeId = "vol-1234567890abcdef0"
)

}
