% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{attach_vpn_gateway}
\alias{attach_vpn_gateway}
\title{Attaches a virtual private gateway to a VPC}
\usage{
attach_vpn_gateway(VpcId, VpnGatewayId, DryRun = NULL)
}
\arguments{
\item{VpcId}{The ID of the VPC.}

\item{VpnGatewayId}{The ID of the virtual private gateway.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Attaches a virtual private gateway to a VPC. You can attach one virtual private gateway to one VPC at a time.
}
\details{
For more information, see \href{http://docs.aws.amazon.com/AmazonVPC/latest/UserGuide/VPC_VPN.html}{AWS Managed VPN Connections} in the \emph{Amazon Virtual Private Cloud User Guide}.
}
\examples{

}
