% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{cancel_export_task}
\alias{cancel_export_task}
\title{Cancels an active export task}
\usage{
cancel_export_task(ExportTaskId)
}
\arguments{
\item{ExportTaskId}{The ID of the export task. This is the ID returned by \code{CreateInstanceExportTask}.}
}
\description{
Cancels an active export task. The request removes all artifacts of the export, including any partially-created Amazon S3 objects. If the export task is complete or is in the process of transferring the final disk image, the command fails and returns an error.
}
\examples{

}
