% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{cancel_reserved_instances_listing}
\alias{cancel_reserved_instances_listing}
\title{Cancels the specified Reserved Instance listing in the Reserved Instance Marketplace}
\usage{
cancel_reserved_instances_listing(
  ReservedInstancesListingId = "string"
)
}
\arguments{
\item{ReservedInstancesListingId}{[required] The ID of the Reserved Instance listing.}
}
\description{
Cancels the specified Reserved Instance listing in the Reserved Instance Marketplace.
}
\details{
For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ri-market-general.html}{Reserved Instance Marketplace} in the \emph{Amazon Elastic Compute Cloud User Guide}.
}
\examples{

}
