% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{cancel_spot_fleet_requests}
\alias{cancel_spot_fleet_requests}
\title{Cancels the specified Spot Fleet requests}
\usage{
cancel_spot_fleet_requests(SpotFleetRequestIds, TerminateInstances,
  DryRun = NULL)
}
\arguments{
\item{SpotFleetRequestIds}{The IDs of the Spot Fleet requests.}

\item{TerminateInstances}{Indicates whether to terminate instances for a Spot Fleet request if it is canceled successfully.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Cancels the specified Spot Fleet requests.
}
\details{
After you cancel a Spot Fleet request, the Spot Fleet launches no new Spot Instances. You must specify whether the Spot Fleet should also terminate its Spot Instances. If you terminate the instances, the Spot Fleet request enters the \code{cancelled_terminating} state. Otherwise, the Spot Fleet request enters the \code{cancelled_running} state and the instances continue to run until they are interrupted or you terminate them manually.
}
\examples{
# This example cancels the specified Spot fleet request and terminates its
# associated Spot Instances.
cancel_spot_fleet_requests(
  SpotFleetRequestIds = list(
    "sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE"
  ),
  TerminateInstances = TRUE
)

# This example cancels the specified Spot fleet request without
# terminating its associated Spot Instances.
cancel_spot_fleet_requests(
  SpotFleetRequestIds = list(
    "sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE"
  ),
  TerminateInstances = FALSE
)

}
