% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{cancel_spot_instance_requests}
\alias{cancel_spot_instance_requests}
\title{Cancels one or more Spot Instance requests}
\usage{
cancel_spot_instance_requests(SpotInstanceRequestIds, DryRun = NULL)
}
\arguments{
\item{SpotInstanceRequestIds}{One or more Spot Instance request IDs.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Cancels one or more Spot Instance requests.
}
\details{
Canceling a Spot Instance request does not terminate running Spot Instances associated with the request.
}
\examples{
# This example cancels a Spot Instance request.
cancel_spot_instance_requests(
  SpotInstanceRequestIds = list(
    "sir-08b93456"
  )
)

}
