% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{confirm_product_instance}
\alias{confirm_product_instance}
\title{Determines whether a product code is associated with an instance}
\usage{
confirm_product_instance(
  InstanceId = "string",
  ProductCode = "string",
  DryRun = TRUE|FALSE
)
}
\arguments{
\item{InstanceId}{[required] The ID of the instance.}

\item{ProductCode}{[required] The product code. This must be a product code that you own.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Determines whether a product code is associated with an instance. This action can only be used by the owner of the product code. It is useful when a product code owner must verify whether another user's instance is eligible for support.
}
\examples{
# This example determines whether the specified product code is associated
# with the specified instance.
confirm_product_instance(
  InstanceId = "i-1234567890abcdef0",
  ProductCode = "774F4FF8"
)

}
