% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{copy_fpga_image}
\alias{copy_fpga_image}
\title{Copies the specified Amazon FPGA Image (AFI) to the current region}
\usage{
copy_fpga_image(SourceFpgaImageId, SourceRegion, DryRun = NULL,
  Description = NULL, Name = NULL, ClientToken = NULL)
}
\arguments{
\item{SourceFpgaImageId}{The ID of the source AFI.}

\item{SourceRegion}{The region that contains the source AFI.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{Description}{The description for the new AFI.}

\item{Name}{The name for the new AFI. The default is the name of the source AFI.}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/Run_Instance_Idempotency.html}{Ensuring Idempotency}.}
}
\description{
Copies the specified Amazon FPGA Image (AFI) to the current region.
}
\examples{

}
