% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{create_egress_only_internet_gateway}
\alias{create_egress_only_internet_gateway}
\title{[IPv6 only] Creates an egress-only internet gateway for your VPC}
\usage{
create_egress_only_internet_gateway(
  ClientToken = "string",
  DryRun = TRUE|FALSE,
  VpcId = "string"
)
}
\arguments{
\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/Run_Instance_Idempotency.html}{How to Ensure Idempotency}.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{VpcId}{[required] The ID of the VPC for which to create the egress-only internet gateway.}
}
\description{
[IPv6 only] Creates an egress-only internet gateway for your VPC. An egress-only internet gateway is used to enable outbound communication over IPv6 from instances in your VPC to the internet, and prevents hosts outside of your VPC from initiating an IPv6 connection with your instance.
}
\examples{

}
