% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{create_fleet}
\alias{create_fleet}
\title{Launches an EC2 Fleet}
\usage{
create_fleet(LaunchTemplateConfigs, TargetCapacitySpecification,
  DryRun = NULL, ClientToken = NULL, SpotOptions = NULL,
  OnDemandOptions = NULL, ExcessCapacityTerminationPolicy = NULL,
  TerminateInstancesWithExpiration = NULL, Type = NULL, ValidFrom = NULL,
  ValidUntil = NULL, ReplaceUnhealthyInstances = NULL,
  TagSpecifications = NULL)
}
\arguments{
\item{LaunchTemplateConfigs}{The configuration for the EC2 Fleet.}

\item{TargetCapacitySpecification}{The \code{TotalTargetCapacity}, \code{OnDemandTargetCapacity}, \code{SpotTargetCapacity}, and \code{DefaultCapacityType} structure.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{ClientToken}{Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/Run_Instance_Idempotency.html}{Ensuring Idempotency}.}

\item{SpotOptions}{Describes the configuration of Spot Instances in an EC2 Fleet.}

\item{OnDemandOptions}{The allocation strategy of On-Demand Instances in an EC2 Fleet.}

\item{ExcessCapacityTerminationPolicy}{Indicates whether running instances should be terminated if the total target capacity of the EC2 Fleet is decreased below the current size of the EC2 Fleet.}

\item{TerminateInstancesWithExpiration}{Indicates whether running instances should be terminated when the EC2 Fleet expires.}

\item{Type}{The type of request. Indicates whether the EC2 Fleet only \code{requests} the target capacity, or also attempts to \code{maintain} it. If you request a certain target capacity, EC2 Fleet only places the required requests. It does not attempt to replenish instances if capacity is diminished, and does not submit requests in alternative capacity pools if capacity is unavailable. To maintain a certain target capacity, EC2 Fleet places the required requests to meet this target capacity. It also automatically replenishes any interrupted Spot Instances. Default: \code{maintain}.}

\item{ValidFrom}{The start date and time of the request, in UTC format (for example, \emph{YYYY}-\emph{MM}-\emph{DD}T\emph{HH}:\emph{MM}:\emph{SS}Z). The default is to start fulfilling the request immediately.}

\item{ValidUntil}{The end date and time of the request, in UTC format (for example, \emph{YYYY}-\emph{MM}-\emph{DD}T\emph{HH}:\emph{MM}:\emph{SS}Z). At this point, no new EC2 Fleet requests are placed or able to fulfill the request. The default end date is 7 days from the current date.}

\item{ReplaceUnhealthyInstances}{Indicates whether EC2 Fleet should replace unhealthy instances.}

\item{TagSpecifications}{The key-value pair for tagging the EC2 Fleet request on creation. The value for \code{ResourceType} must be \code{fleet}, otherwise the fleet request fails. To tag instances at launch, specify the tags in the \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-launch-templates.html#create-launch-template}{launch template}. For information about tagging after launch, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/Using_Tags.html#tag-resources}{Tagging Your Resources}.}
}
\description{
Launches an EC2 Fleet.
}
\details{
You can create a single EC2 Fleet that includes multiple launch specifications that vary by instance type, AMI, Availability Zone, or subnet.

For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-fleet.html}{Launching an EC2 Fleet} in the \emph{Amazon Elastic Compute Cloud User Guide}.
}
\examples{

}
