% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{create_fpga_image}
\alias{create_fpga_image}
\title{Creates an Amazon FPGA Image (AFI) from the specified design checkpoint (DCP)}
\usage{
create_fpga_image(InputStorageLocation, DryRun = NULL,
  LogsStorageLocation = NULL, Description = NULL, Name = NULL,
  ClientToken = NULL)
}
\arguments{
\item{InputStorageLocation}{The location of the encrypted design checkpoint in Amazon S3. The input must be a tarball.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{LogsStorageLocation}{The location in Amazon S3 for the output logs.}

\item{Description}{A description for the AFI.}

\item{Name}{A name for the AFI.}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/Run_Instance_Idempotency.html}{Ensuring Idempotency}.}
}
\description{
Creates an Amazon FPGA Image (AFI) from the specified design checkpoint (DCP).
}
\details{
The create operation is asynchronous. To verify that the AFI is ready for use, check the output logs.

An AFI contains the FPGA bitstream that is ready to download to an FPGA. You can securely deploy an AFI on one or more FPGA-accelerated instances. For more information, see the \href{https://github.com/aws/aws-fpga/}{AWS FPGA Hardware Development Kit}.
}
\examples{

}
