% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{create_launch_template}
\alias{create_launch_template}
\title{Creates a launch template}
\usage{
create_launch_template(LaunchTemplateName, LaunchTemplateData, DryRun = NULL,
  ClientToken = NULL, VersionDescription = NULL)
}
\arguments{
\item{LaunchTemplateName}{A name for the launch template.}

\item{LaunchTemplateData}{The information for the launch template.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{ClientToken}{Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/Run_Instance_Idempotency.html}{Ensuring Idempotency}.}

\item{VersionDescription}{A description for the first version of the launch template.}
}
\description{
Creates a launch template. A launch template contains the parameters to launch an instance. When you launch an instance using RunInstances, you can specify a launch template instead of providing the launch parameters in the request.
}
\examples{
# This example creates a launch template that specifies the subnet in
# which to launch the instance, assigns a public IP address and an IPv6
# address to the instance, and creates a tag for the instance.
create_launch_template(
  LaunchTemplateData = list(
    ImageId = "ami-8c1be5f6",
    InstanceType = "t2.small",
    NetworkInterfaces = list(
      list(
        AssociatePublicIpAddress = TRUE,
        DeviceIndex = 0L,
        Ipv6AddressCount = 1L,
        SubnetId = "subnet-7b16de0c"
      )
    ),
    TagSpecifications = list(
      list(
        ResourceType = "instance",
        Tags = list(
          list(
            Key = "Name",
            Value = "webserver"
          )
        )
      )
    )
  ),
  LaunchTemplateName = "my-template",
  VersionDescription = "WebVersion1"
)

}
