% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{create_launch_template_version}
\alias{create_launch_template_version}
\title{Creates a new version for a launch template}
\usage{
create_launch_template_version(LaunchTemplateData, DryRun = NULL,
  ClientToken = NULL, LaunchTemplateId = NULL, LaunchTemplateName = NULL,
  SourceVersion = NULL, VersionDescription = NULL)
}
\arguments{
\item{LaunchTemplateData}{The information for the launch template.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{ClientToken}{Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/Run_Instance_Idempotency.html}{Ensuring Idempotency}.}

\item{LaunchTemplateId}{The ID of the launch template. You must specify either the launch template ID or launch template name in the request.}

\item{LaunchTemplateName}{The name of the launch template. You must specify either the launch template ID or launch template name in the request.}

\item{SourceVersion}{The version number of the launch template version on which to base the new version. The new version inherits the same launch parameters as the source version, except for parameters that you specify in LaunchTemplateData.}

\item{VersionDescription}{A description for the version of the launch template.}
}
\description{
Creates a new version for a launch template. You can specify an existing version of launch template from which to base the new version.
}
\details{
Launch template versions are numbered in the order in which they are created. You cannot specify, change, or replace the numbering of launch template versions.
}
\examples{
# This example creates a new launch template version based on version 1 of
# the specified launch template and specifies a different AMI ID.
create_launch_template_version(
  LaunchTemplateData = list(
    ImageId = "ami-c998b6b2"
  ),
  LaunchTemplateId = "lt-0abcd290751193123",
  SourceVersion = "1",
  VersionDescription = "WebVersion2"
)

}
