% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{create_network_interface}
\alias{create_network_interface}
\title{Creates a network interface in the specified subnet}
\usage{
create_network_interface(SubnetId, Description = NULL, DryRun = NULL,
  Groups = NULL, Ipv6AddressCount = NULL, Ipv6Addresses = NULL,
  PrivateIpAddress = NULL, PrivateIpAddresses = NULL,
  SecondaryPrivateIpAddressCount = NULL)
}
\arguments{
\item{SubnetId}{The ID of the subnet to associate with the network interface.}

\item{Description}{A description for the network interface.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{Groups}{The IDs of one or more security groups.}

\item{Ipv6AddressCount}{The number of IPv6 addresses to assign to a network interface. Amazon EC2 automatically selects the IPv6 addresses from the subnet range. You can't use this option if specifying specific IPv6 addresses. If your subnet has the \code{AssignIpv6AddressOnCreation} attribute set to \code{true}, you can specify \code{0} to override this setting.}

\item{Ipv6Addresses}{One or more specific IPv6 addresses from the IPv6 CIDR block range of your subnet. You can't use this option if you're specifying a number of IPv6 addresses.}

\item{PrivateIpAddress}{The primary private IPv4 address of the network interface. If you don't specify an IPv4 address, Amazon EC2 selects one for you from the subnet's IPv4 CIDR range. If you specify an IP address, you cannot indicate any IP addresses specified in \code{privateIpAddresses} as primary (only one IP address can be designated as primary).}

\item{PrivateIpAddresses}{One or more private IPv4 addresses.}

\item{SecondaryPrivateIpAddressCount}{The number of secondary private IPv4 addresses to assign to a network interface. When you specify a number of secondary IPv4 addresses, Amazon EC2 selects these IP addresses within the subnet's IPv4 CIDR range. You can't specify this option and specify more than one private IP address using \code{privateIpAddresses}.

The number of IP addresses you can assign to a network interface varies by instance type. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/using-eni.html#AvailableIpPerENI}{IP Addresses Per ENI Per Instance Type} in the \emph{Amazon Virtual Private Cloud User Guide}.}
}
\description{
Creates a network interface in the specified subnet.
}
\details{
For more information about network interfaces, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/using-eni.html}{Elastic Network Interfaces} in the \emph{Amazon Virtual Private Cloud User Guide}.
}
\examples{
# This example creates a network interface for the specified subnet.
create_network_interface(
  Description = "my network interface",
  Groups = list(
    "sg-903004f8"
  ),
  PrivateIpAddress = "10.0.2.17",
  SubnetId = "subnet-9d4a7b6c"
)

}
