% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{create_subnet}
\alias{create_subnet}
\title{Creates a subnet in an existing VPC}
\usage{
create_subnet(
  AvailabilityZone = "string",
  AvailabilityZoneId = "string",
  CidrBlock = "string",
  Ipv6CidrBlock = "string",
  VpcId = "string",
  DryRun = TRUE|FALSE
)
}
\arguments{
\item{AvailabilityZone}{The Availability Zone for the subnet.

Default: AWS selects one for you. If you create more than one subnet in your VPC, we may not necessarily select a different zone for each subnet.}

\item{AvailabilityZoneId}{The AZ ID of the subnet.}

\item{CidrBlock}{[required] The IPv4 network range for the subnet, in CIDR notation. For example, \code{10.0.0.0/24}.}

\item{Ipv6CidrBlock}{The IPv6 network range for the subnet, in CIDR notation. The subnet size must use a /64 prefix length.}

\item{VpcId}{[required] The ID of the VPC.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Creates a subnet in an existing VPC.
}
\details{
When you create each subnet, you provide the VPC ID and IPv4 CIDR block for the subnet. After you create a subnet, you can't change its CIDR block. The size of the subnet's IPv4 CIDR block can be the same as a VPC's IPv4 CIDR block, or a subset of a VPC's IPv4 CIDR block. If you create more than one subnet in a VPC, the subnets' CIDR blocks must not overlap. The smallest IPv4 subnet (and VPC) you can create uses a /28 netmask (16 IPv4 addresses), and the largest uses a /16 netmask (65,536 IPv4 addresses).

If you've associated an IPv6 CIDR block with your VPC, you can create a subnet with an IPv6 CIDR block that uses a /64 prefix length.

AWS reserves both the first four and the last IPv4 address in each subnet's CIDR block. They're not available for use.

If you add more than one subnet to a VPC, they're set up in a star topology with a logical router in the middle.

If you launch an instance in a VPC using an Amazon EBS-backed AMI, the IP address doesn't change if you stop and restart the instance (unlike a similar instance launched outside a VPC, which gets a new IP address when restarted). It's therefore possible to have a subnet with no running instances (they're all stopped), but no remaining IP addresses available.

For more information about subnets, see \href{http://docs.aws.amazon.com/AmazonVPC/latest/UserGuide/VPC_Subnets.html}{Your VPC and Subnets} in the \emph{Amazon Virtual Private Cloud User Guide}.
}
\examples{
# This example creates a subnet in the specified VPC with the specified
# CIDR block. We recommend that you let us select an Availability Zone for
# you.
create_subnet(
  CidrBlock = "10.0.1.0/24",
  VpcId = "vpc-a01106c2"
)

}
