% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{create_tags}
\alias{create_tags}
\title{Adds or overwrites one or more tags for the specified Amazon EC2 resource or resources}
\usage{
create_tags(Resources, Tags, DryRun = NULL)
}
\arguments{
\item{Resources}{The IDs of one or more resources to tag. For example, ami-1a2b3c4d.}

\item{Tags}{One or more tags. The \code{value} parameter is required, but if you don't want the tag to have a value, specify the parameter with no value, and we set the value to an empty string.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Adds or overwrites one or more tags for the specified Amazon EC2 resource or resources. Each resource can have a maximum of 50 tags. Each tag consists of a key and optional value. Tag keys must be unique per resource.
}
\details{
For more information about tags, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/Using_Tags.html}{Tagging Your Resources} in the \emph{Amazon Elastic Compute Cloud User Guide}. For more information about creating IAM policies that control users' access to resources based on tags, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-supported-iam-actions-resources.html}{Supported Resource-Level Permissions for Amazon EC2 API Actions} in the \emph{Amazon Elastic Compute Cloud User Guide}.
}
\examples{
# This example adds the tag Stack=production to the specified image, or
# overwrites an existing tag for the AMI where the tag key is Stack.
create_tags(
  Resources = list(
    "ami-78a54011"
  ),
  Tags = list(
    list(
      Key = "Stack",
      Value = "production"
    )
  )
)

}
