% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{create_vpn_connection}
\alias{create_vpn_connection}
\title{Creates a VPN connection between an existing virtual private gateway and a VPN customer gateway}
\usage{
create_vpn_connection(
  CustomerGatewayId = "string",
  Type = "string",
  VpnGatewayId = "string",
  TransitGatewayId = "string",
  DryRun = TRUE|FALSE,
  Options = list(
    StaticRoutesOnly = TRUE|FALSE,
    TunnelOptions = list(
      list(
        TunnelInsideCidr = "string",
        PreSharedKey = "string"
      )
    )
  )
)
}
\arguments{
\item{CustomerGatewayId}{[required] The ID of the customer gateway.}

\item{Type}{[required] The type of VPN connection (\code{ipsec.1}).}

\item{VpnGatewayId}{The ID of the virtual private gateway. If you specify a virtual private gateway, you cannot specify a transit gateway.}

\item{TransitGatewayId}{The ID of the transit gateway. If you specify a transit gateway, you cannot specify a virtual private gateway.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{Options}{The options for the VPN connection.}
}
\description{
Creates a VPN connection between an existing virtual private gateway and a VPN customer gateway. The only supported connection type is \code{ipsec.1}.
}
\details{
The response includes information that you need to give to your network administrator to configure your customer gateway.

We strongly recommend that you use HTTPS when calling this operation because the response contains sensitive cryptographic information for configuring your customer gateway.

If you decide to shut down your VPN connection for any reason and later create a new VPN connection, you must reconfigure your customer gateway with the new information returned from this call.

This is an idempotent operation. If you perform the operation more than once, Amazon EC2 doesn't return an error.

For more information, see \href{http://docs.aws.amazon.com/AmazonVPC/latest/UserGuide/VPC_VPN.html}{AWS Managed VPN Connections} in the \emph{Amazon Virtual Private Cloud User Guide}.
}
\examples{

}
