% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{delete_customer_gateway}
\alias{delete_customer_gateway}
\title{Deletes the specified customer gateway}
\usage{
delete_customer_gateway(
  CustomerGatewayId = "string",
  DryRun = TRUE|FALSE
)
}
\arguments{
\item{CustomerGatewayId}{[required] The ID of the customer gateway.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified customer gateway. You must delete the VPN connection before you can delete the customer gateway.
}
\examples{
# This example deletes the specified customer gateway.
delete_customer_gateway(
  CustomerGatewayId = "cgw-0e11f167"
)

}
