% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{delete_fleets}
\alias{delete_fleets}
\title{Deletes the specified EC2 Fleet}
\usage{
delete_fleets(FleetIds, TerminateInstances, DryRun = NULL)
}
\arguments{
\item{FleetIds}{The IDs of the EC2 Fleets.}

\item{TerminateInstances}{Indicates whether to terminate instances for an EC2 Fleet if it is deleted successfully.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified EC2 Fleet.
}
\details{
After you delete an EC2 Fleet, it launches no new instances. You must specify whether an EC2 Fleet should also terminate its instances. If you terminate the instances, the EC2 Fleet enters the \code{deleted_terminating} state. Otherwise, the EC2 Fleet enters the \code{deleted_running} state, and the instances continue to run until they are interrupted or you terminate them manually.
}
\examples{

}
