% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{delete_fpga_image}
\alias{delete_fpga_image}
\title{Deletes the specified Amazon FPGA Image (AFI)}
\usage{
delete_fpga_image(FpgaImageId, DryRun = NULL)
}
\arguments{
\item{FpgaImageId}{The ID of the AFI.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified Amazon FPGA Image (AFI).
}
\examples{

}
