% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{delete_internet_gateway}
\alias{delete_internet_gateway}
\title{Deletes the specified internet gateway}
\usage{
delete_internet_gateway(
  DryRun = TRUE|FALSE,
  InternetGatewayId = "string"
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{InternetGatewayId}{[required] The ID of the internet gateway.}
}
\description{
Deletes the specified internet gateway. You must detach the internet gateway from the VPC before you can delete it.
}
\examples{
# This example deletes the specified Internet gateway.
delete_internet_gateway(
  InternetGatewayId = "igw-c0a643a9"
)

}
