% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{delete_key_pair}
\alias{delete_key_pair}
\title{Deletes the specified key pair, by removing the public key from Amazon EC2}
\usage{
delete_key_pair(
  KeyName = "string",
  DryRun = TRUE|FALSE
)
}
\arguments{
\item{KeyName}{[required] The name of the key pair.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified key pair, by removing the public key from Amazon EC2.
}
\examples{
# This example deletes the specified key pair.
delete_key_pair(
  KeyName = "my-key-pair"
)

}
