% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{delete_launch_template}
\alias{delete_launch_template}
\title{Deletes a launch template}
\usage{
delete_launch_template(DryRun = NULL, LaunchTemplateId = NULL,
  LaunchTemplateName = NULL)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{LaunchTemplateId}{The ID of the launch template. You must specify either the launch template ID or launch template name in the request.}

\item{LaunchTemplateName}{The name of the launch template. You must specify either the launch template ID or launch template name in the request.}
}
\description{
Deletes a launch template. Deleting a launch template deletes all of its versions.
}
\examples{
# This example deletes the specified launch template.
delete_launch_template(
  LaunchTemplateId = "lt-0abcd290751193123"
)

}
