% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{delete_network_acl}
\alias{delete_network_acl}
\title{Deletes the specified network ACL}
\usage{
delete_network_acl(NetworkAclId, DryRun = NULL)
}
\arguments{
\item{NetworkAclId}{The ID of the network ACL.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified network ACL. You can't delete the ACL if it's associated with any subnets. You can't delete the default network ACL.
}
\examples{
# This example deletes the specified network ACL.
delete_network_acl(
  NetworkAclId = "acl-5fb85d36"
)

}
