% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{delete_network_acl_entry}
\alias{delete_network_acl_entry}
\title{Deletes the specified ingress or egress entry (rule) from the specified network ACL}
\usage{
delete_network_acl_entry(Egress, NetworkAclId, RuleNumber, DryRun = NULL)
}
\arguments{
\item{Egress}{Indicates whether the rule is an egress rule.}

\item{NetworkAclId}{The ID of the network ACL.}

\item{RuleNumber}{The rule number of the entry to delete.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified ingress or egress entry (rule) from the specified network ACL.
}
\examples{
# This example deletes ingress rule number 100 from the specified network
# ACL.
delete_network_acl_entry(
  Egress = TRUE,
  NetworkAclId = "acl-5fb85d36",
  RuleNumber = 100L
)

}
