% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{delete_placement_group}
\alias{delete_placement_group}
\title{Deletes the specified placement group}
\usage{
delete_placement_group(GroupName, DryRun = NULL)
}
\arguments{
\item{GroupName}{The name of the placement group.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified placement group. You must terminate all instances in the placement group before you can delete the placement group. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/placement-groups.html}{Placement Groups} in the \emph{Amazon Elastic Compute Cloud User Guide}.
}
\examples{
# This example deletes the specified placement group.
# 
delete_placement_group(
  GroupName = "my-cluster"
)

}
