% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{delete_route}
\alias{delete_route}
\title{Deletes the specified route from the specified route table}
\usage{
delete_route(RouteTableId, DestinationCidrBlock = NULL,
  DestinationIpv6CidrBlock = NULL, DryRun = NULL)
}
\arguments{
\item{RouteTableId}{The ID of the route table.}

\item{DestinationCidrBlock}{The IPv4 CIDR range for the route. The value you specify must match the CIDR for the route exactly.}

\item{DestinationIpv6CidrBlock}{The IPv6 CIDR range for the route. The value you specify must match the CIDR for the route exactly.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified route from the specified route table.
}
\examples{
# This example deletes the specified route from the specified route table.
delete_route(
  DestinationCidrBlock = "0.0.0.0/0",
  RouteTableId = "rtb-22574640"
)

}
