% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{delete_spot_datafeed_subscription}
\alias{delete_spot_datafeed_subscription}
\title{Deletes the data feed for Spot Instances}
\usage{
delete_spot_datafeed_subscription(
  DryRun = TRUE|FALSE
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the data feed for Spot Instances.
}
\examples{
# This example deletes a Spot data feed subscription for the account.
delete_spot_datafeed_subscription()

}
