% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{delete_subnet}
\alias{delete_subnet}
\title{Deletes the specified subnet}
\usage{
delete_subnet(
  SubnetId = "string",
  DryRun = TRUE|FALSE
)
}
\arguments{
\item{SubnetId}{[required] The ID of the subnet.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified subnet. You must terminate all running instances in the subnet before you can delete the subnet.
}
\examples{
# This example deletes the specified subnet.
delete_subnet(
  SubnetId = "subnet-9d4a7b6c"
)

}
