% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{delete_vpc}
\alias{delete_vpc}
\title{Deletes the specified VPC}
\usage{
delete_vpc(VpcId, DryRun = NULL)
}
\arguments{
\item{VpcId}{The ID of the VPC.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified VPC. You must detach or delete all gateways and resources that are associated with the VPC before you can delete it. For example, you must terminate all instances running in the VPC, delete all security groups associated with the VPC (except the default one), delete all route tables associated with the VPC (except the default one), and so on.
}
\examples{
# This example deletes the specified VPC.
delete_vpc(
  VpcId = "vpc-a01106c2"
)

}
