% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{delete_vpc_endpoints}
\alias{delete_vpc_endpoints}
\title{Deletes one or more specified VPC endpoints}
\usage{
delete_vpc_endpoints(VpcEndpointIds, DryRun = NULL)
}
\arguments{
\item{VpcEndpointIds}{One or more VPC endpoint IDs.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes one or more specified VPC endpoints. Deleting a gateway endpoint also deletes the endpoint routes in the route tables that were associated with the endpoint. Deleting an interface endpoint deletes the endpoint network interfaces.
}
\examples{

}
