% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{delete_vpn_connection_route}
\alias{delete_vpn_connection_route}
\title{Deletes the specified static route associated with a VPN connection between an existing virtual private gateway and a VPN customer gateway}
\usage{
delete_vpn_connection_route(DestinationCidrBlock, VpnConnectionId)
}
\arguments{
\item{DestinationCidrBlock}{The CIDR block associated with the local subnet of the customer network.}

\item{VpnConnectionId}{The ID of the VPN connection.}
}
\description{
Deletes the specified static route associated with a VPN connection between an existing virtual private gateway and a VPN customer gateway. The static route allows traffic to be routed from the virtual private gateway to the VPN customer gateway.
}
\examples{

}
