% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{delete_vpn_gateway}
\alias{delete_vpn_gateway}
\title{Deletes the specified virtual private gateway}
\usage{
delete_vpn_gateway(VpnGatewayId, DryRun = NULL)
}
\arguments{
\item{VpnGatewayId}{[required] The ID of the virtual private gateway.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes the specified virtual private gateway. We recommend that before you delete a virtual private gateway, you detach it from the VPC and delete the VPN connection. Note that you don't need to delete the virtual private gateway if you plan to delete and recreate the VPN connection between your VPC and your network.
}
\section{Accepted Parameters}{
\preformatted{delete_vpn_gateway(
  VpnGatewayId = "string",
  DryRun = TRUE|FALSE
)
}
}

