% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{describe_dhcp_options}
\alias{describe_dhcp_options}
\title{Describes one or more of your DHCP options sets}
\usage{
describe_dhcp_options(DhcpOptionsIds = NULL, Filters = NULL,
  DryRun = NULL)
}
\arguments{
\item{DhcpOptionsIds}{The IDs of one or more DHCP options sets.

Default: Describes all your DHCP options sets.}

\item{Filters}{One or more filters.
\itemize{
\item \code{dhcp-options-id} - The ID of a set of DHCP options.
\item \code{key} - The key for one of the options (for example, \code{domain-name}).
\item \code{value} - The value for one of the options.
\item \code{tag}:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key \code{Owner} and the value \code{TeamA}, specify \code{tag:Owner} for the filter name and \code{TeamA} for the filter value.
\item \code{tag-key} - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.
}}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describes one or more of your DHCP options sets.
}
\details{
For more information, see \href{http://docs.aws.amazon.com/AmazonVPC/latest/UserGuide/VPC_DHCP_Options.html}{DHCP Options Sets} in the \emph{Amazon Virtual Private Cloud User Guide}.
}
\section{Accepted Parameters}{
\preformatted{describe_dhcp_options(
  DhcpOptionsIds = list(
    "string"
  ),
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  DryRun = TRUE|FALSE
)
}
}

\examples{
# This example describes the specified DHCP options set.
\donttest{describe_dhcp_options(
  DhcpOptionsIds = list(
    "dopt-d9070ebb"
  )
)}

}
