% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{describe_egress_only_internet_gateways}
\alias{describe_egress_only_internet_gateways}
\title{Describes one or more of your egress-only internet gateways}
\usage{
describe_egress_only_internet_gateways(DryRun = NULL,
  EgressOnlyInternetGatewayIds = NULL, MaxResults = NULL,
  NextToken = NULL)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{EgressOnlyInternetGatewayIds}{One or more egress-only internet gateway IDs.}

\item{MaxResults}{The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned \code{NextToken} value. This value can be between 5 and 1000. If \code{MaxResults} is given a value larger than 1000, only 1000 results are returned.}

\item{NextToken}{The token to retrieve the next page of results.}
}
\description{
Describes one or more of your egress-only internet gateways.
}
\section{Accepted Parameters}{
\preformatted{describe_egress_only_internet_gateways(
  DryRun = TRUE|FALSE,
  EgressOnlyInternetGatewayIds = list(
    "string"
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

