% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{describe_export_tasks}
\alias{describe_export_tasks}
\title{Describes one or more of your export tasks}
\usage{
describe_export_tasks(ExportTaskIds = NULL)
}
\arguments{
\item{ExportTaskIds}{One or more export task IDs.}
}
\description{
Describes one or more of your export tasks.
}
\section{Accepted Parameters}{
\preformatted{describe_export_tasks(
  ExportTaskIds = list(
    "string"
  )
)
}
}

