% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{describe_fpga_image_attribute}
\alias{describe_fpga_image_attribute}
\title{Describes the specified attribute of the specified Amazon FPGA Image (AFI)}
\usage{
describe_fpga_image_attribute(FpgaImageId, Attribute, DryRun = NULL)
}
\arguments{
\item{FpgaImageId}{The ID of the AFI.}

\item{Attribute}{The AFI attribute.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describes the specified attribute of the specified Amazon FPGA Image (AFI).
}
\examples{

}
