% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{describe_hosts}
\alias{describe_hosts}
\title{Describes one or more of your Dedicated Hosts}
\usage{
describe_hosts(Filter = NULL, HostIds = NULL, MaxResults = NULL,
  NextToken = NULL)
}
\arguments{
\item{Filter}{One or more filters.
\itemize{
\item \code{auto-placement} - Whether auto-placement is enabled or disabled (\code{on} | \code{off}).
\item \code{availability-zone} - The Availability Zone of the host.
\item \code{client-token} - The idempotency token that you provided when you allocated the host.
\item \code{host-reservation-id} - The ID of the reservation assigned to this host.
\item \code{instance-type} - The instance type size that the Dedicated Host is configured to support.
\item \code{state} - The allocation state of the Dedicated Host (\code{available} | \code{under-assessment} | \code{permanent-failure} | \code{released} | \code{released-permanent-failure}).
\item \code{tag-key} - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.
}}

\item{HostIds}{The IDs of the Dedicated Hosts. The IDs are used for targeted instance launches.}

\item{MaxResults}{The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned \code{nextToken} value. This value can be between 5 and 500. If \code{maxResults} is given a larger value than 500, you receive an error. You cannot specify this parameter and the host IDs parameter in the same request.}

\item{NextToken}{The token to retrieve the next page of results.}
}
\description{
Describes one or more of your Dedicated Hosts.
}
\details{
The results describe only the Dedicated Hosts in the Region you're currently using. All listed instances consume capacity on your Dedicated Host. Dedicated Hosts that have recently been released are listed with the state \code{released}.
}
\section{Accepted Parameters}{
\preformatted{describe_hosts(
  Filter = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  HostIds = list(
    "string"
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

