% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{describe_iam_instance_profile_associations}
\alias{describe_iam_instance_profile_associations}
\title{Describes your IAM instance profile associations}
\usage{
describe_iam_instance_profile_associations(
  AssociationIds = list(
    "string"
  ),
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
\arguments{
\item{AssociationIds}{One or more IAM instance profile associations.}

\item{Filters}{One or more filters.
\itemize{
\item \code{instance-id} - The ID of the instance.
\item \code{state} - The state of the association (\code{associating} | \code{associated} | \code{disassociating} | \code{disassociated}).
}}

\item{MaxResults}{The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned \code{NextToken} value.}

\item{NextToken}{The token to request the next page of results.}
}
\description{
Describes your IAM instance profile associations.
}
\examples{
# This example describes the specified IAM instance profile association.
describe_iam_instance_profile_associations(
  AssociationIds = list(
    "iip-assoc-0db249b1f25fa24b8"
  )
)

}
