% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{describe_images}
\alias{describe_images}
\title{Describes one or more of the images (AMIs, AKIs, and ARIs) available to you}
\usage{
describe_images(
  ExecutableUsers = list(
    "string"
  ),
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  ImageIds = list(
    "string"
  ),
  Owners = list(
    "string"
  ),
  DryRun = TRUE|FALSE
)
}
\arguments{
\item{ExecutableUsers}{Scopes the images by users with explicit launch permissions. Specify an AWS account ID, \code{self} (the sender of the request), or \code{all} (public AMIs).}

\item{Filters}{One or more filters.
\itemize{
\item \code{architecture} - The image architecture (\code{i386} | \code{x86_64}).
\item \code{block-device-mapping.delete-on-termination} - A Boolean value that indicates whether the Amazon EBS volume is deleted on instance termination.
\item \code{block-device-mapping.device-name} - The device name specified in the block device mapping (for example, \code{/dev/sdh} or \code{xvdh}).
\item \code{block-device-mapping.snapshot-id} - The ID of the snapshot used for the EBS volume.
\item \code{block-device-mapping.volume-size} - The volume size of the EBS volume, in GiB.
\item \code{block-device-mapping.volume-type} - The volume type of the EBS volume (\code{gp2} | \code{io1} | \code{st1 }| \code{sc1} | \code{standard}).
\item \code{description} - The description of the image (provided during image creation).
\item \code{ena-support} - A Boolean that indicates whether enhanced networking with ENA is enabled.
\item \code{hypervisor} - The hypervisor type (\code{ovm} | \code{xen}).
\item \code{image-id} - The ID of the image.
\item \code{image-type} - The image type (\code{machine} | \code{kernel} | \code{ramdisk}).
\item \code{is-public} - A Boolean that indicates whether the image is public.
\item \code{kernel-id} - The kernel ID.
\item \code{manifest-location} - The location of the image manifest.
\item \code{name} - The name of the AMI (provided during image creation).
\item \code{owner-alias} - String value from an Amazon-maintained list (\code{amazon} | \code{aws-marketplace} | \code{microsoft}) of snapshot owners. Not to be confused with the user-configured AWS account alias, which is set from the IAM console.
\item \code{owner-id} - The AWS account ID of the image owner.
\item \code{platform} - The platform. To only list Windows-based AMIs, use \code{windows}.
\item \code{product-code} - The product code.
\item \code{product-code.type} - The type of the product code (\code{devpay} | \code{marketplace}).
\item \code{ramdisk-id} - The RAM disk ID.
\item \code{root-device-name} - The device name of the root device volume (for example, \code{/dev/sda1}).
\item \code{root-device-type} - The type of the root device volume (\code{ebs} | \code{instance-store}).
\item \code{state} - The state of the image (\code{available} | \code{pending} | \code{failed}).
\item \code{state-reason-code} - The reason code for the state change.
\item \code{state-reason-message} - The message for the state change.
\item \code{sriov-net-support} - A value of \code{simple} indicates that enhanced networking with the Intel 82599 VF interface is enabled.
\item \code{tag}:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key \code{Owner} and the value \code{TeamA}, specify \code{tag:Owner} for the filter name and \code{TeamA} for the filter value.
\item \code{tag-key} - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.
\item \code{virtualization-type} - The virtualization type (\code{paravirtual} | \code{hvm}).
}}

\item{ImageIds}{One or more image IDs.

Default: Describes all images available to you.}

\item{Owners}{Filters the images by the owner. Specify an AWS account ID, \code{self} (owner is the sender of the request), or an AWS owner alias (valid values are \code{amazon} | \code{aws-marketplace} | \code{microsoft}). Omitting this option returns all images for which you have launch permissions, regardless of ownership.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describes one or more of the images (AMIs, AKIs, and ARIs) available to you. Images available to you include public images, private images that you own, and private images owned by other AWS accounts but for which you have explicit launch permissions.
}
\details{
Deregistered images are included in the returned results for an unspecified interval after deregistration.
}
\examples{
# This example describes the specified AMI.
describe_images(
  ImageIds = list(
    "ami-5731123e"
  )
)

}
