% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{describe_instance_attribute}
\alias{describe_instance_attribute}
\title{Describes the specified attribute of the specified instance}
\usage{
describe_instance_attribute(Attribute, InstanceId, DryRun = NULL)
}
\arguments{
\item{Attribute}{The instance attribute.

Note: The \code{enaSupport} attribute is not supported at this time.}

\item{InstanceId}{The ID of the instance.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describes the specified attribute of the specified instance. You can specify only one attribute at a time. Valid attribute values are: \code{instanceType} | \code{kernel} | \code{ramdisk} | \code{userData} | \code{disableApiTermination} | \code{instanceInitiatedShutdownBehavior} | \code{rootDeviceName} | \code{blockDeviceMapping} | \code{productCodes} | \code{sourceDestCheck} | \code{groupSet} | \code{ebsOptimized} | \code{sriovNetSupport}
}
\examples{
# This example describes the instance type of the specified instance.
# 
describe_instance_attribute(
  Attribute = "instanceType",
  InstanceId = "i-1234567890abcdef0"
)

# This example describes the ``disableApiTermination`` attribute of the
# specified instance.
# 
describe_instance_attribute(
  Attribute = "disableApiTermination",
  InstanceId = "i-1234567890abcdef0"
)

# This example describes the ``blockDeviceMapping`` attribute of the
# specified instance.
# 
describe_instance_attribute(
  Attribute = "blockDeviceMapping",
  InstanceId = "i-1234567890abcdef0"
)

}
