% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{describe_network_interface_permissions}
\alias{describe_network_interface_permissions}
\title{Describes the permissions for your network interfaces}
\usage{

  describe_network_interface_permissions(NetworkInterfacePermissionIds = NULL,
  Filters = NULL, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NetworkInterfacePermissionIds}{One or more network interface permission IDs.}

\item{Filters}{One or more filters.
\itemize{
\item \code{network-interface-permission.network-interface-permission-id} - The ID of the permission.
\item \code{network-interface-permission.network-interface-id} - The ID of the network interface.
\item \code{network-interface-permission.aws-account-id} - The AWS account ID.
\item \code{network-interface-permission.aws-service} - The AWS service.
\item \code{network-interface-permission.permission} - The type of permission (\code{INSTANCE-ATTACH} | \code{EIP-ASSOCIATE}).
}}

\item{NextToken}{The token to request the next page of results.}

\item{MaxResults}{The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned \code{NextToken} value. If this parameter is not specified, up to 50 results are returned by default.}
}
\description{
Describes the permissions for your network interfaces.
}
\section{Accepted Parameters}{
\preformatted{describe_network_interface_permissions(
  NetworkInterfacePermissionIds = list(
    "string"
  ),
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  NextToken = "string",
  MaxResults = 123
)
}
}

