% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{describe_network_interfaces}
\alias{describe_network_interfaces}
\title{Describes one or more of your network interfaces}
\usage{
describe_network_interfaces(
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  DryRun = TRUE|FALSE,
  NetworkInterfaceIds = list(
    "string"
  ),
  NextToken = "string",
  MaxResults = 123
)
}
\arguments{
\item{Filters}{One or more filters.
\itemize{
\item \code{addresses.private-ip-address} - The private IPv4 addresses associated with the network interface.
\item \code{addresses.primary} - Whether the private IPv4 address is the primary IP address associated with the network interface.
\item \code{addresses.association.public-ip} - The association ID returned when the network interface was associated with the Elastic IP address (IPv4).
\item \code{addresses.association.owner-id} - The owner ID of the addresses associated with the network interface.
\item \code{association.association-id} - The association ID returned when the network interface was associated with an IPv4 address.
\item \code{association.allocation-id} - The allocation ID returned when you allocated the Elastic IP address (IPv4) for your network interface.
\item \code{association.ip-owner-id} - The owner of the Elastic IP address (IPv4) associated with the network interface.
\item \code{association.public-ip} - The address of the Elastic IP address (IPv4) bound to the network interface.
\item \code{association.public-dns-name} - The public DNS name for the network interface (IPv4).
\item \code{attachment.attachment-id} - The ID of the interface attachment.
\item \code{attachment.attach.time} - The time that the network interface was attached to an instance.
\item \code{attachment.delete-on-termination} - Indicates whether the attachment is deleted when an instance is terminated.
\item \code{attachment.device-index} - The device index to which the network interface is attached.
\item \code{attachment.instance-id} - The ID of the instance to which the network interface is attached.
\item \code{attachment.instance-owner-id} - The owner ID of the instance to which the network interface is attached.
\item \code{attachment.nat-gateway-id} - The ID of the NAT gateway to which the network interface is attached.
\item \code{attachment.status} - The status of the attachment (\code{attaching} | \code{attached} | \code{detaching} | \code{detached}).
\item \code{availability-zone} - The Availability Zone of the network interface.
\item \code{description} - The description of the network interface.
\item \code{group-id} - The ID of a security group associated with the network interface.
\item \code{group-name} - The name of a security group associated with the network interface.
\item \code{ipv6-addresses.ipv6-address} - An IPv6 address associated with the network interface.
\item \code{mac-address} - The MAC address of the network interface.
\item \code{network-interface-id} - The ID of the network interface.
\item \code{owner-id} - The AWS account ID of the network interface owner.
\item \code{private-ip-address} - The private IPv4 address or addresses of the network interface.
\item \code{private-dns-name} - The private DNS name of the network interface (IPv4).
\item \code{requester-id} - The ID of the entity that launched the instance on your behalf (for example, AWS Management Console, Auto Scaling, and so on).
\item \code{requester-managed} - Indicates whether the network interface is being managed by an AWS service (for example, AWS Management Console, Auto Scaling, and so on).
\item \code{source-desk-check} - Indicates whether the network interface performs source/destination checking. A value of \code{true} means checking is enabled, and \code{false} means checking is disabled. The value must be \code{false} for the network interface to perform network address translation (NAT) in your VPC.
\item \code{status} - The status of the network interface. If the network interface is not attached to an instance, the status is \code{available}; if a network interface is attached to an instance the status is \code{in-use}.
\item \code{subnet-id} - The ID of the subnet for the network interface.
\item \code{tag}:<key> - The key/value combination of a tag assigned to the resource. Use the tag key in the filter name and the tag value as the filter value. For example, to find all resources that have a tag with the key \code{Owner} and the value \code{TeamA}, specify \code{tag:Owner} for the filter name and \code{TeamA} for the filter value.
\item \code{tag-key} - The key of a tag assigned to the resource. Use this filter to find all resources assigned a tag with a specific key, regardless of the tag value.
\item \code{vpc-id} - The ID of the VPC for the network interface.
}}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{NetworkInterfaceIds}{One or more network interface IDs.

Default: Describes all your network interfaces.}

\item{NextToken}{The token to retrieve the next page of results.}

\item{MaxResults}{The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results.}
}
\description{
Describes one or more of your network interfaces.
}
\examples{
# 
describe_network_interfaces(
  NetworkInterfaceIds = list(
    "eni-e5aa89a3"
  )
)

}
