% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{describe_prefix_lists}
\alias{describe_prefix_lists}
\title{Describes available AWS services in a prefix list format, which includes the prefix list name and prefix list ID of the service and the IP address range for the service}
\usage{
describe_prefix_lists(DryRun = NULL, Filters = NULL, MaxResults = NULL,
  NextToken = NULL, PrefixListIds = NULL)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{Filters}{One or more filters.
\itemize{
\item \code{prefix-list-id}: The ID of a prefix list.
\item \code{prefix-list-name}: The name of a prefix list.
}}

\item{MaxResults}{The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results.

Constraint: If the value specified is greater than 1000, we return only 1000 items.}

\item{NextToken}{The token for the next set of items to return. (You received this token from a prior call.)}

\item{PrefixListIds}{One or more prefix list IDs.}
}
\description{
Describes available AWS services in a prefix list format, which includes the prefix list name and prefix list ID of the service and the IP address range for the service. A prefix list ID is required for creating an outbound security group rule that allows traffic from a VPC to access an AWS service through a gateway VPC endpoint. Currently, the services that support this action are Amazon S3 and Amazon DynamoDB.
}
\examples{

}
