% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{describe_reserved_instances_offerings}
\alias{describe_reserved_instances_offerings}
\title{Describes Reserved Instance offerings that are available for purchase}
\usage{
describe_reserved_instances_offerings(AvailabilityZone = NULL,
  Filters = NULL, IncludeMarketplace = NULL, InstanceType = NULL,
  MaxDuration = NULL, MaxInstanceCount = NULL, MinDuration = NULL,
  OfferingClass = NULL, ProductDescription = NULL,
  ReservedInstancesOfferingIds = NULL, DryRun = NULL,
  InstanceTenancy = NULL, MaxResults = NULL, NextToken = NULL,
  OfferingType = NULL)
}
\arguments{
\item{AvailabilityZone}{The Availability Zone in which the Reserved Instance can be used.}

\item{Filters}{One or more filters.
\itemize{
\item \code{availability-zone} - The Availability Zone where the Reserved Instance can be used.
\item \code{duration} - The duration of the Reserved Instance (for example, one year or three years), in seconds (\code{31536000} | \code{94608000}).
\item \code{fixed-price} - The purchase price of the Reserved Instance (for example, 9800.0).
\item \code{instance-type} - The instance type that is covered by the reservation.
\item \code{marketplace} - Set to \code{true} to show only Reserved Instance Marketplace offerings. When this filter is not used, which is the default behavior, all offerings from both AWS and the Reserved Instance Marketplace are listed.
\item \code{product-description} - The Reserved Instance product platform description. Instances that include \code{(Amazon VPC)} in the product platform description will only be displayed to EC2-Classic account holders and are for use with Amazon VPC. (\code{Linux/UNIX} | \code{Linux/UNIX (Amazon VPC)} | \code{SUSE Linux} | \code{SUSE Linux (Amazon VPC)} | \code{Red Hat Enterprise Linux} | \code{Red Hat Enterprise Linux (Amazon VPC)} | \code{Windows} | \code{Windows (Amazon VPC)} | \code{Windows with SQL Server Standard} | \code{Windows with SQL Server Standard (Amazon VPC)} | \code{Windows with SQL Server Web} | \code{ Windows with SQL Server Web (Amazon VPC)} | \code{Windows with SQL Server Enterprise} | \code{Windows with SQL Server Enterprise (Amazon VPC)})
\item \code{reserved-instances-offering-id} - The Reserved Instances offering ID.
\item \code{scope} - The scope of the Reserved Instance (\code{Availability Zone} or \code{Region}).
\item \code{usage-price} - The usage price of the Reserved Instance, per hour (for example, 0.84).
}}

\item{IncludeMarketplace}{Include Reserved Instance Marketplace offerings in the response.}

\item{InstanceType}{The instance type that the reservation will cover (for example, \code{m1.small}). For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/instance-types.html}{Instance Types} in the \emph{Amazon Elastic Compute Cloud User Guide}.}

\item{MaxDuration}{The maximum duration (in seconds) to filter when searching for offerings.

Default: 94608000 (3 years)}

\item{MaxInstanceCount}{The maximum number of instances to filter when searching for offerings.

Default: 20}

\item{MinDuration}{The minimum duration (in seconds) to filter when searching for offerings.

Default: 2592000 (1 month)}

\item{OfferingClass}{The offering class of the Reserved Instance. Can be \code{standard} or \code{convertible}.}

\item{ProductDescription}{The Reserved Instance product platform description. Instances that include \code{(Amazon VPC)} in the description are for use with Amazon VPC.}

\item{ReservedInstancesOfferingIds}{One or more Reserved Instances offering IDs.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{InstanceTenancy}{The tenancy of the instances covered by the reservation. A Reserved Instance with a tenancy of \code{dedicated} is applied to instances that run in a VPC on single-tenant hardware (i.e., Dedicated Instances).

\strong{Important:} The \code{host} value cannot be used with this parameter. Use the \code{default} or \code{dedicated} values only.

Default: \code{default}}

\item{MaxResults}{The maximum number of results to return for the request in a single page. The remaining results of the initial request can be seen by sending another request with the returned \code{NextToken} value. The maximum is 100.

Default: 100}

\item{NextToken}{The token to retrieve the next page of results.}

\item{OfferingType}{The Reserved Instance offering type. If you are using tools that predate the 2011-11-01 API version, you only have access to the \code{Medium Utilization} Reserved Instance offering type.}
}
\description{
Describes Reserved Instance offerings that are available for purchase. With Reserved Instances, you purchase the right to launch instances for a period of time. During that time period, you do not receive insufficient capacity errors, and you pay a lower usage rate than the rate charged for On-Demand instances for the actual time used.
}
\details{
If you have listed your own Reserved Instances for sale in the Reserved Instance Marketplace, they will be excluded from these results. This is to ensure that you do not purchase your own Reserved Instances.

For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ri-market-general.html}{Reserved Instance Marketplace} in the \emph{Amazon Elastic Compute Cloud User Guide}.
}
\examples{

}
