% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{describe_tags}
\alias{describe_tags}
\title{Describes one or more of the tags for your EC2 resources}
\usage{
describe_tags(DryRun = NULL, Filters = NULL, MaxResults = NULL,
  NextToken = NULL)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{Filters}{One or more filters.
\itemize{
\item \code{key} - The tag key.
\item \code{resource-id} - The ID of the resource.
\item \code{resource-type} - The resource type (\code{customer-gateway} | \code{dedicated-host} | \code{dhcp-options} | \code{elastic-ip} | \code{fleet} | \code{fpga-image} | \code{image} | \code{instance} | \code{internet-gateway} | \code{launch-template} | \code{natgateway} | \code{network-acl} | \code{network-interface} | \code{reserved-instances} | \code{route-table} | \code{security-group} | \code{snapshot} | \code{spot-instances-request} | \code{subnet} | \code{volume} | \code{vpc} | \code{vpc-peering-connection} | \code{vpn-connection} | \code{vpn-gateway}).
\item \code{tag}:<key> - The key/value combination of the tag. For example, specify "tag:Owner" for the filter name and "TeamA" for the filter value to find resources with the tag "Owner=TeamA".
\item \code{value} - The tag value.
}}

\item{MaxResults}{The maximum number of results to return in a single call. This value can be between 5 and 1000. To retrieve the remaining results, make another call with the returned \code{NextToken} value.}

\item{NextToken}{The token to retrieve the next page of results.}
}
\description{
Describes one or more of the tags for your EC2 resources.
}
\details{
For more information about tags, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/Using_Tags.html}{Tagging Your Resources} in the \emph{Amazon Elastic Compute Cloud User Guide}.
}
\examples{
# This example describes the tags for the specified instance.
describe_tags(
  Filters = list(
    list(
      Name = "resource-id",
      Values = list(
        "i-1234567890abcdef8"
      )
    )
  )
)

}
