% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{detach_internet_gateway}
\alias{detach_internet_gateway}
\title{Detaches an internet gateway from a VPC, disabling connectivity between the internet and the VPC}
\usage{
detach_internet_gateway(InternetGatewayId, VpcId, DryRun = NULL)
}
\arguments{
\item{InternetGatewayId}{The ID of the internet gateway.}

\item{VpcId}{The ID of the VPC.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Detaches an internet gateway from a VPC, disabling connectivity between the internet and the VPC. The VPC must not contain any running instances with Elastic IP addresses or public IPv4 addresses.
}
\examples{
# This example detaches the specified Internet gateway from the specified
# VPC.
detach_internet_gateway(
  InternetGatewayId = "igw-c0a643a9",
  VpcId = "vpc-a01106c2"
)

}
