% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{disable_transit_gateway_route_table_propagation}
\alias{disable_transit_gateway_route_table_propagation}
\title{Disables the specified resource attachment from propagating routes to the specified propagation route table}
\usage{

  disable_transit_gateway_route_table_propagation(TransitGatewayRouteTableId,
  TransitGatewayAttachmentId, DryRun = NULL)
}
\arguments{
\item{TransitGatewayRouteTableId}{[required] The ID of the propagation route table.}

\item{TransitGatewayAttachmentId}{[required] The ID of the attachment.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Disables the specified resource attachment from propagating routes to the specified propagation route table.
}
\section{Accepted Parameters}{
\preformatted{disable_transit_gateway_route_table_propagation(
  TransitGatewayRouteTableId = "string",
  TransitGatewayAttachmentId = "string",
  DryRun = TRUE|FALSE
)
}
}

