% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{disable_vgw_route_propagation}
\alias{disable_vgw_route_propagation}
\title{Disables a virtual private gateway (VGW) from propagating routes to a specified route table of a VPC}
\usage{
disable_vgw_route_propagation(
  GatewayId = "string",
  RouteTableId = "string"
)
}
\arguments{
\item{GatewayId}{[required] The ID of the virtual private gateway.}

\item{RouteTableId}{[required] The ID of the route table.}
}
\description{
Disables a virtual private gateway (VGW) from propagating routes to a specified route table of a VPC.
}
\examples{
# This example disables the specified virtual private gateway from
# propagating static routes to the specified route table.
disable_vgw_route_propagation(
  GatewayId = "vgw-9a4cacf3",
  RouteTableId = "rtb-22574640"
)

}
