% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{disassociate_iam_instance_profile}
\alias{disassociate_iam_instance_profile}
\title{Disassociates an IAM instance profile from a running or stopped instance}
\usage{
disassociate_iam_instance_profile(AssociationId)
}
\arguments{
\item{AssociationId}{The ID of the IAM instance profile association.}
}
\description{
Disassociates an IAM instance profile from a running or stopped instance.
}
\details{
Use DescribeIamInstanceProfileAssociations to get the association ID.
}
\examples{
# This example disassociates the specified IAM instance profile from an
# instance.
disassociate_iam_instance_profile(
  AssociationId = "iip-assoc-05020b59952902f5f"
)

}
