% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{enable_volume_io}
\alias{enable_volume_io}
\title{Enables I/O operations for a volume that had I/O operations disabled because the data on the volume was potentially inconsistent}
\usage{
enable_volume_io(
  DryRun = TRUE|FALSE,
  VolumeId = "string"
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{VolumeId}{[required] The ID of the volume.}
}
\description{
Enables I/O operations for a volume that had I/O operations disabled because the data on the volume was potentially inconsistent.
}
\examples{
# This example enables I/O on volume ``vol-1234567890abcdef0``.
enable_volume_io(
  VolumeId = "vol-1234567890abcdef0"
)

}
