% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{export_client_vpn_client_certificate_revocation_list}
\alias{export_client_vpn_client_certificate_revocation_list}
\title{Downloads the client certificate revocation list for the specified Client VPN endpoint}
\usage{
export_client_vpn_client_certificate_revocation_list(
  ClientVpnEndpointId = "string",
  DryRun = TRUE|FALSE
)
}
\arguments{
\item{ClientVpnEndpointId}{[required] The ID of the Client VPN endpoint.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Downloads the client certificate revocation list for the specified Client VPN endpoint.
}
\examples{

}
