% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{export_transit_gateway_routes}
\alias{export_transit_gateway_routes}
\title{Exports routes from the specified transit gateway route table to the specified S3 bucket}
\usage{
export_transit_gateway_routes(
  TransitGatewayRouteTableId = "string",
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  S3Bucket = "string",
  DryRun = TRUE|FALSE
)
}
\arguments{
\item{TransitGatewayRouteTableId}{[required] The ID of the route table.}

\item{Filters}{One or more filters. The possible values are:
\itemize{
\item \code{transit-gateway-route-destination-cidr-block} - The CIDR range.
\item \code{transit-gateway-route-state} - The state of the route (\code{active} | \code{blackhole}).
\item \code{transit-gateway-route-transit-gateway-attachment-id} - The ID of the attachment.
\item \code{transit-gateway-route-type} - The route type (\code{static} | \code{propagated}).
\item \code{transit-gateway-route-vpn-connection-id} - The ID of the VPN connection.
}}

\item{S3Bucket}{[required] The name of the S3 bucket.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Exports routes from the specified transit gateway route table to the specified S3 bucket. By default, all routes are exported. Alternatively, you can filter by CIDR range.
}
\examples{

}
