% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{import_image}
\alias{import_image}
\title{Import single or multi-volume disk images or EBS snapshots into an Amazon Machine Image (AMI)}
\usage{
import_image(Architecture = NULL, ClientData = NULL,
  ClientToken = NULL, Description = NULL, DiskContainers = NULL,
  DryRun = NULL, Hypervisor = NULL, LicenseType = NULL,
  Platform = NULL, RoleName = NULL)
}
\arguments{
\item{Architecture}{The architecture of the virtual machine.

Valid values: \code{i386} | \code{x86_64}}

\item{ClientData}{The client-specific data.}

\item{ClientToken}{The token to enable idempotency for VM import requests.}

\item{Description}{A description string for the import image task.}

\item{DiskContainers}{Information about the disk containers.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{Hypervisor}{The target hypervisor platform.

Valid values: \code{xen}}

\item{LicenseType}{The license type to be used for the Amazon Machine Image (AMI) after importing.

\strong{Note:} You may only use BYOL if you have existing licenses with rights to use these licenses in a third party cloud like AWS. For more information, see \href{http://docs.aws.amazon.com/vm-import/latest/userguide/vmimport-image-import.html#prerequisites-image}{Prerequisites} in the VM Import/Export User Guide.

Valid values: \code{AWS} | \code{BYOL}}

\item{Platform}{The operating system of the virtual machine.

Valid values: \code{Windows} | \code{Linux}}

\item{RoleName}{The name of the role to use when not using the default role, 'vmimport'.}
}
\description{
Import single or multi-volume disk images or EBS snapshots into an Amazon Machine Image (AMI). For more information, see \href{http://docs.aws.amazon.com/vm-import/latest/userguide/vmimport-image-import.html}{Importing a VM as an Image Using VM Import/Export} in the \emph{VM Import/Export User Guide}.
}
\section{Accepted Parameters}{
\preformatted{import_image(
  Architecture = "string",
  ClientData = list(
    Comment = "string",
    UploadEnd = as.POSIXct("2015-01-01"),
    UploadSize = 123.0,
    UploadStart = as.POSIXct("2015-01-01")
  ),
  ClientToken = "string",
  Description = "string",
  DiskContainers = list(
    list(
      Description = "string",
      DeviceName = "string",
      Format = "string",
      SnapshotId = "string",
      Url = "string",
      UserBucket = list(
        S3Bucket = "string",
        S3Key = "string"
      )
    )
  ),
  DryRun = TRUE|FALSE,
  Hypervisor = "string",
  LicenseType = "string",
  Platform = "string",
  RoleName = "string"
)
}
}

