% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{import_snapshot}
\alias{import_snapshot}
\title{Imports a disk into an EBS snapshot}
\usage{
import_snapshot(ClientData = NULL, ClientToken = NULL, Description = NULL,
  DiskContainer = NULL, DryRun = NULL, RoleName = NULL)
}
\arguments{
\item{ClientData}{The client-specific data.}

\item{ClientToken}{Token to enable idempotency for VM import requests.}

\item{Description}{The description string for the import snapshot task.}

\item{DiskContainer}{Information about the disk container.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{RoleName}{The name of the role to use when not using the default role, 'vmimport'.}
}
\description{
Imports a disk into an EBS snapshot.
}
\examples{

}
