% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{modify_reserved_instances}
\alias{modify_reserved_instances}
\title{Modifies the Availability Zone, instance count, instance type, or network platform (EC2-Classic or EC2-VPC) of your Reserved Instances}
\usage{
modify_reserved_instances(ReservedInstancesIds, TargetConfigurations,
  ClientToken = NULL)
}
\arguments{
\item{ReservedInstancesIds}{The IDs of the Reserved Instances to modify.}

\item{TargetConfigurations}{The configuration settings for the Reserved Instances to modify.}

\item{ClientToken}{A unique, case-sensitive token you provide to ensure idempotency of your modification request. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/Run_Instance_Idempotency.html}{Ensuring Idempotency}.}
}
\description{
Modifies the Availability Zone, instance count, instance type, or network platform (EC2-Classic or EC2-VPC) of your Reserved Instances. The Reserved Instances to be modified must be identical, except for Availability Zone, network platform, and instance type.
}
\details{
For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ri-modifying.html}{Modifying Reserved Instances} in the Amazon Elastic Compute Cloud User Guide.
}
\examples{

}
