% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{modify_vpc_endpoint_connection_notification}
\alias{modify_vpc_endpoint_connection_notification}
\title{Modifies a connection notification for VPC endpoint or VPC endpoint service}
\usage{
modify_vpc_endpoint_connection_notification(
  DryRun = TRUE|FALSE,
  ConnectionNotificationId = "string",
  ConnectionNotificationArn = "string",
  ConnectionEvents = list(
    "string"
  )
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{ConnectionNotificationId}{[required] The ID of the notification.}

\item{ConnectionNotificationArn}{The ARN for the SNS topic for the notification.}

\item{ConnectionEvents}{One or more events for the endpoint. Valid values are \code{Accept}, \code{Connect}, \code{Delete}, and \code{Reject}.}
}
\description{
Modifies a connection notification for VPC endpoint or VPC endpoint service. You can change the SNS topic for the notification, or the events for which to be notified.
}
\examples{

}
