% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{monitor_instances}
\alias{monitor_instances}
\title{Enables detailed monitoring for a running instance}
\usage{
monitor_instances(InstanceIds, DryRun = NULL)
}
\arguments{
\item{InstanceIds}{One or more instance IDs.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Enables detailed monitoring for a running instance. Otherwise, basic monitoring is enabled. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/using-cloudwatch.html}{Monitoring Your Instances and Volumes} in the \emph{Amazon Elastic Compute Cloud User Guide}.
}
\details{
To disable detailed monitoring, see .
}
\examples{

}
