% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{move_address_to_vpc}
\alias{move_address_to_vpc}
\title{Moves an Elastic IP address from the EC2-Classic platform to the EC2-VPC platform}
\usage{
move_address_to_vpc(PublicIp, DryRun = NULL)
}
\arguments{
\item{PublicIp}{The Elastic IP address.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Moves an Elastic IP address from the EC2-Classic platform to the EC2-VPC platform. The Elastic IP address must be allocated to your account for more than 24 hours, and it must not be associated with an instance. After the Elastic IP address is moved, it is no longer available for use in the EC2-Classic platform, unless you move it back using the RestoreAddressToClassic request. You cannot move an Elastic IP address that was originally allocated for use in the EC2-VPC platform to the EC2-Classic platform.
}
\examples{
# This example moves the specified Elastic IP address to the EC2-VPC
# platform.
move_address_to_vpc(
  PublicIp = "54.123.4.56"
)

}
