% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_package.R
\docType{package}
\name{paws.ec2-package}
\alias{paws.ec2}
\alias{paws.ec2-package}
\title{paws.ec2: Amazon Elastic Compute Cloud}
\description{
Amazon Elastic Compute Cloud (Amazon EC2) provides
    secure and resizable computing capacity in the AWS cloud. Using Amazon
    EC2 eliminates the need to invest in hardware up front, so you can
    develop and deploy applications faster.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
