% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{purchase_reserved_instances_offering}
\alias{purchase_reserved_instances_offering}
\title{Purchases a Reserved Instance for use with your account}
\usage{
purchase_reserved_instances_offering(InstanceCount, ReservedInstancesOfferingId,
  DryRun = NULL, LimitPrice = NULL)
}
\arguments{
\item{InstanceCount}{The number of Reserved Instances to purchase.}

\item{ReservedInstancesOfferingId}{The ID of the Reserved Instance offering to purchase.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{LimitPrice}{Specified for Reserved Instance Marketplace offerings to limit the total order and ensure that the Reserved Instances are not purchased at unexpected prices.}
}
\description{
Purchases a Reserved Instance for use with your account. With Reserved Instances, you pay a lower hourly rate compared to On-Demand instance pricing.
}
\details{
Use DescribeReservedInstancesOfferings to get a list of Reserved Instance offerings that match your specifications. After you've purchased a Reserved Instance, you can check for your new Reserved Instance with DescribeReservedInstances.

For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/concepts-on-demand-reserved-instances.html}{Reserved Instances} and \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ri-market-general.html}{Reserved Instance Marketplace} in the \emph{Amazon Elastic Compute Cloud User Guide}.
}
\examples{

}
