% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{register_image}
\alias{register_image}
\title{Registers an AMI}
\usage{
register_image(ImageLocation = NULL, Architecture = NULL,
  BlockDeviceMappings = NULL, Description = NULL, DryRun = NULL,
  EnaSupport = NULL, KernelId = NULL, Name, BillingProducts = NULL,
  RamdiskId = NULL, RootDeviceName = NULL, SriovNetSupport = NULL,
  VirtualizationType = NULL)
}
\arguments{
\item{ImageLocation}{The full path to your AMI manifest in Amazon S3 storage.}

\item{Architecture}{The architecture of the AMI.

Default: For Amazon EBS-backed AMIs, \code{i386}. For instance store-backed AMIs, the architecture specified in the manifest file.}

\item{BlockDeviceMappings}{One or more block device mapping entries.}

\item{Description}{A description for your AMI.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{EnaSupport}{Set to \code{true} to enable enhanced networking with ENA for the AMI and any instances that you launch from the AMI.

This option is supported only for HVM AMIs. Specifying this option with a PV AMI can make instances launched from the AMI unreachable.}

\item{KernelId}{The ID of the kernel.}

\item{Name}{[required] A name for your AMI.

Constraints: 3-128 alphanumeric characters, parentheses (()), square brackets ([]), spaces ( ), periods (.), slashes (/), dashes (-), single quotes ('), at-signs (@), or underscores(\_)}

\item{BillingProducts}{The billing product codes. Your account must be authorized to specify billing product codes. Otherwise, you can use the AWS Marketplace to bill for the use of an AMI.}

\item{RamdiskId}{The ID of the RAM disk.}

\item{RootDeviceName}{The device name of the root device volume (for example, \code{/dev/sda1}).}

\item{SriovNetSupport}{Set to \code{simple} to enable enhanced networking with the Intel 82599 Virtual Function interface for the AMI and any instances that you launch from the AMI.

There is no way to disable \code{sriovNetSupport} at this time.

This option is supported only for HVM AMIs. Specifying this option with a PV AMI can make instances launched from the AMI unreachable.}

\item{VirtualizationType}{The type of virtualization (\code{hvm} | \code{paravirtual}).

Default: \code{paravirtual}}
}
\description{
Registers an AMI. When you're creating an AMI, this is the final step you must complete before you can launch an instance from the AMI. For more information about creating AMIs, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/creating-an-ami.html}{Creating Your Own AMIs} in the \emph{Amazon Elastic Compute Cloud User Guide}.
}
\details{
For Amazon EBS-backed instances, CreateImage creates and registers the AMI in a single request, so you don't have to register the AMI yourself.

You can also use \code{RegisterImage} to create an Amazon EBS-backed Linux AMI from a snapshot of a root device volume. You specify the snapshot using the block device mapping. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/instance-launch-snapshot.html}{Launching a Linux Instance from a Backup} in the \emph{Amazon Elastic Compute Cloud User Guide}.

You can't register an image where a secondary (non-root) snapshot has AWS Marketplace product codes.

Some Linux distributions, such as Red Hat Enterprise Linux (RHEL) and SUSE Linux Enterprise Server (SLES), use the EC2 billing product code associated with an AMI to verify the subscription status for package updates. Creating an AMI from an EBS snapshot does not maintain this billing code, and instances launched from such an AMI are not able to connect to package update infrastructure. If you purchase a Reserved Instance offering for one of these Linux distributions and launch instances using an AMI that does not contain the required billing code, your Reserved Instance is not applied to these instances.

To create an AMI for operating systems that require a billing code, see CreateImage.

If needed, you can deregister an AMI at any time. Any modifications you make to an AMI backed by an instance store volume invalidates its registration. If you make changes to an image, deregister the previous image and register the new image.
}
\section{Accepted Parameters}{
\preformatted{register_image(
  ImageLocation = "string",
  Architecture = "i386"|"x86_64",
  BlockDeviceMappings = list(
    list(
      DeviceName = "string",
      VirtualName = "string",
      Ebs = list(
        DeleteOnTermination = TRUE|FALSE,
        Iops = 123,
        SnapshotId = "string",
        VolumeSize = 123,
        VolumeType = "standard"|"io1"|"gp2"|"sc1"|"st1",
        Encrypted = TRUE|FALSE,
        KmsKeyId = "string"
      ),
      NoDevice = "string"
    )
  ),
  Description = "string",
  DryRun = TRUE|FALSE,
  EnaSupport = TRUE|FALSE,
  KernelId = "string",
  Name = "string",
  BillingProducts = list(
    "string"
  ),
  RamdiskId = "string",
  RootDeviceName = "string",
  SriovNetSupport = "string",
  VirtualizationType = "string"
)
}
}

