% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{release_hosts}
\alias{release_hosts}
\title{When you no longer want to use an On-Demand Dedicated Host it can be released}
\usage{
release_hosts(
  HostIds = list(
    "string"
  )
)
}
\arguments{
\item{HostIds}{[required] The IDs of the Dedicated Hosts to release.}
}
\description{
When you no longer want to use an On-Demand Dedicated Host it can be released. On-Demand billing is stopped and the host goes into \code{released} state. The host ID of Dedicated Hosts that have been released can no longer be specified in another request, for example, to modify the host. You must stop or terminate all instances on a host before it can be released.
}
\details{
When Dedicated Hosts are released, it may take some time for them to stop counting toward your limit and you may receive capacity errors when trying to allocate new Dedicated Hosts. Wait a few minutes and then try again.

Released hosts still appear in a DescribeHosts response.
}
\examples{

}
