% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{replace_iam_instance_profile_association}
\alias{replace_iam_instance_profile_association}
\title{Replaces an IAM instance profile for the specified running instance}
\usage{
replace_iam_instance_profile_association(IamInstanceProfile, AssociationId)
}
\arguments{
\item{IamInstanceProfile}{The IAM instance profile.}

\item{AssociationId}{The ID of the existing IAM instance profile association.}
}
\description{
Replaces an IAM instance profile for the specified running instance. You can use this action to change the IAM instance profile that's associated with an instance without having to disassociate the existing IAM instance profile first.
}
\details{
Use DescribeIamInstanceProfileAssociations to get the association ID.
}
\examples{

}
