% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{replace_network_acl_association}
\alias{replace_network_acl_association}
\title{Changes which network ACL a subnet is associated with}
\usage{
replace_network_acl_association(AssociationId, NetworkAclId, DryRun = NULL)
}
\arguments{
\item{AssociationId}{The ID of the current association between the original network ACL and the subnet.}

\item{NetworkAclId}{The ID of the new network ACL to associate with the subnet.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Changes which network ACL a subnet is associated with. By default when you create a subnet, it's automatically associated with the default network ACL. For more information, see \href{http://docs.aws.amazon.com/AmazonVPC/latest/UserGuide/VPC_ACLs.html}{Network ACLs} in the \emph{Amazon Virtual Private Cloud User Guide}.
}
\details{
This is an idempotent operation.
}
\examples{
# This example associates the specified network ACL with the subnet for
# the specified network ACL association.
replace_network_acl_association(
  AssociationId = "aclassoc-e5b95c8c",
  NetworkAclId = "acl-5fb85d36"
)

}
