% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{replace_network_acl_entry}
\alias{replace_network_acl_entry}
\title{Replaces an entry (rule) in a network ACL}
\usage{
replace_network_acl_entry(Egress, NetworkAclId, Protocol, RuleAction,
  RuleNumber, CidrBlock = NULL, DryRun = NULL, IcmpTypeCode = NULL,
  Ipv6CidrBlock = NULL, PortRange = NULL)
}
\arguments{
\item{Egress}{Indicates whether to replace the egress rule.

Default: If no value is specified, we replace the ingress rule.}

\item{NetworkAclId}{The ID of the ACL.}

\item{Protocol}{The IP protocol. You can specify \code{all} or \code{-1} to mean all protocols. If you specify \code{all}, \code{-1}, or a protocol number other than \code{tcp}, \code{udp}, or \code{icmp}, traffic on all ports is allowed, regardless of any ports or ICMP types or codes you that specify. If you specify protocol \code{58} (ICMPv6) and specify an IPv4 CIDR block, traffic for all ICMP types and codes allowed, regardless of any that you specify. If you specify protocol \code{58} (ICMPv6) and specify an IPv6 CIDR block, you must specify an ICMP type and code.}

\item{RuleAction}{Indicates whether to allow or deny the traffic that matches the rule.}

\item{RuleNumber}{The rule number of the entry to replace.}

\item{CidrBlock}{The IPv4 network range to allow or deny, in CIDR notation (for example \code{172.16.0.0/24}).}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IcmpTypeCode}{ICMP protocol: The ICMP or ICMPv6 type and code. Required if specifying the ICMP (1) protocol, or protocol 58 (ICMPv6) with an IPv6 CIDR block.}

\item{Ipv6CidrBlock}{The IPv6 network range to allow or deny, in CIDR notation (for example \code{2001:bd8:1234:1a00::/64}).}

\item{PortRange}{TCP or UDP protocols: The range of ports the rule applies to. Required if specifying TCP (6) or UDP (17) for the protocol.}
}
\description{
Replaces an entry (rule) in a network ACL. For more information, see \href{http://docs.aws.amazon.com/AmazonVPC/latest/UserGuide/VPC_ACLs.html}{Network ACLs} in the \emph{Amazon Virtual Private Cloud User Guide}.
}
\examples{
# This example replaces an entry for the specified network ACL. The new
# rule 100 allows ingress traffic from 203.0.113.12/24 on UDP port 53
# (DNS) into any associated subnet.
replace_network_acl_entry(
  CidrBlock = "203.0.113.12/24",
  Egress = FALSE,
  NetworkAclId = "acl-5fb85d36",
  PortRange = list(
    From = 53L,
    To = 53L
  ),
  Protocol = "udp",
  RuleAction = "allow",
  RuleNumber = 100L
)

}
