% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{request_spot_instances}
\alias{request_spot_instances}
\title{Creates a Spot Instance request}
\usage{
request_spot_instances(AvailabilityZoneGroup = NULL,
  BlockDurationMinutes = NULL, ClientToken = NULL, DryRun = NULL,
  InstanceCount = NULL, LaunchGroup = NULL, LaunchSpecification = NULL,
  SpotPrice = NULL, Type = NULL, ValidFrom = NULL, ValidUntil = NULL,
  InstanceInterruptionBehavior = NULL)
}
\arguments{
\item{AvailabilityZoneGroup}{The user-specified name for a logical grouping of requests.

When you specify an Availability Zone group in a Spot Instance request, all Spot Instances in the request are launched in the same Availability Zone. Instance proximity is maintained with this parameter, but the choice of Availability Zone is not. The group applies only to requests for Spot Instances of the same instance type. Any additional Spot Instance requests that are specified with the same Availability Zone group name are launched in that same Availability Zone, as long as at least one instance from the group is still active.

If there is no active instance running in the Availability Zone group that you specify for a new Spot Instance request (all instances are terminated, the request is expired, or the maximum price you specified falls below current Spot price), then Amazon EC2 launches the instance in any Availability Zone where the constraint can be met. Consequently, the subsequent set of Spot Instances could be placed in a different zone from the original request, even if you specified the same Availability Zone group.

Default: Instances are launched in any available Availability Zone.}

\item{BlockDurationMinutes}{The required duration for the Spot Instances (also known as Spot blocks), in minutes. This value must be a multiple of 60 (60, 120, 180, 240, 300, or 360).

The duration period starts as soon as your Spot Instance receives its instance ID. At the end of the duration period, Amazon EC2 marks the Spot Instance for termination and provides a Spot Instance termination notice, which gives the instance a two-minute warning before it terminates.

You can't specify an Availability Zone group or a launch group if you specify a duration.}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/Run_Instance_Idempotency.html}{How to Ensure Idempotency} in the \emph{Amazon EC2 User Guide for Linux Instances}.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{InstanceCount}{The maximum number of Spot Instances to launch.

Default: 1}

\item{LaunchGroup}{The instance launch group. Launch groups are Spot Instances that launch together and terminate together.

Default: Instances are launched and terminated individually}

\item{LaunchSpecification}{The launch specification.}

\item{SpotPrice}{The maximum price per hour that you are willing to pay for a Spot Instance. The default is the On-Demand price.}

\item{Type}{The Spot Instance request type.

Default: \code{one-time}}

\item{ValidFrom}{The start date of the request. If this is a one-time request, the request becomes active at this date and time and remains active until all instances launch, the request expires, or the request is canceled. If the request is persistent, the request becomes active at this date and time and remains active until it expires or is canceled.}

\item{ValidUntil}{The end date of the request. If this is a one-time request, the request remains active until all instances launch, the request is canceled, or this date is reached. If the request is persistent, it remains active until it is canceled or this date is reached. The default end date is 7 days from the current date.}

\item{InstanceInterruptionBehavior}{The behavior when a Spot Instance is interrupted. The default is \code{terminate}.}
}
\description{
Creates a Spot Instance request.
}
\details{
For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/spot-requests.html}{Spot Instance Requests} in the \emph{Amazon EC2 User Guide for Linux Instances}.
}
\examples{
# This example creates a one-time Spot Instance request for five instances
# in the specified Availability Zone. If your account supports EC2-VPC
# only, Amazon EC2 launches the instances in the default subnet of the
# specified Availability Zone. If your account supports EC2-Classic,
# Amazon EC2 launches the instances in EC2-Classic in the specified
# Availability Zone.
request_spot_instances(
  InstanceCount = 5L,
  LaunchSpecification = list(
    IamInstanceProfile = list(
      Arn = "arn:aws:iam::123456789012:instance-profile/my-iam-role"
    ),
    ImageId = "ami-1a2b3c4d",
    InstanceType = "m3.medium",
    KeyName = "my-key-pair",
    Placement = list(
      AvailabilityZone = "us-west-2a"
    ),
    SecurityGroupIds = list(
      "sg-1a2b3c4d"
    )
  ),
  SpotPrice = "0.03",
  Type = "one-time"
)

# This example command creates a one-time Spot Instance request for five
# instances in the specified subnet. Amazon EC2 launches the instances in
# the specified subnet. If the VPC is a nondefault VPC, the instances do
# not receive a public IP address by default.
request_spot_instances(
  InstanceCount = 5L,
  LaunchSpecification = list(
    IamInstanceProfile = list(
      Arn = "arn:aws:iam::123456789012:instance-profile/my-iam-role"
    ),
    ImageId = "ami-1a2b3c4d",
    InstanceType = "m3.medium",
    SecurityGroupIds = list(
      "sg-1a2b3c4d"
    ),
    SubnetId = "subnet-1a2b3c4d"
  ),
  SpotPrice = "0.050",
  Type = "one-time"
)

}
