% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{reset_fpga_image_attribute}
\alias{reset_fpga_image_attribute}
\title{Resets the specified attribute of the specified Amazon FPGA Image (AFI) to its default value}
\usage{
reset_fpga_image_attribute(FpgaImageId, DryRun = NULL, Attribute = NULL)
}
\arguments{
\item{FpgaImageId}{The ID of the AFI.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{Attribute}{The attribute.}
}
\description{
Resets the specified attribute of the specified Amazon FPGA Image (AFI) to its default value. You can only reset the load permission attribute.
}
\examples{

}
