% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{reset_image_attribute}
\alias{reset_image_attribute}
\title{Resets an attribute of an AMI to its default value}
\usage{
reset_image_attribute(
  Attribute = "launchPermission",
  ImageId = "string",
  DryRun = TRUE|FALSE
)
}
\arguments{
\item{Attribute}{[required] The attribute to reset (currently you can only reset the launch permission attribute).}

\item{ImageId}{[required] The ID of the AMI.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Resets an attribute of an AMI to its default value.
}
\details{
The productCodes attribute can't be reset.
}
\examples{
# This example resets the launchPermission attribute for the specified
# AMI. By default, AMIs are private.
reset_image_attribute(
  Attribute = "launchPermission",
  ImageId = "ami-5731123e"
)

}
