% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{revoke_security_group_egress}
\alias{revoke_security_group_egress}
\title{[EC2-VPC only] Removes one or more egress rules from a security group for EC2-VPC}
\usage{
revoke_security_group_egress(GroupId, DryRun = NULL, IpPermissions = NULL,
  CidrIp = NULL, FromPort = NULL, IpProtocol = NULL, ToPort = NULL,
  SourceSecurityGroupName = NULL, SourceSecurityGroupOwnerId = NULL)
}
\arguments{
\item{GroupId}{The ID of the security group.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{IpPermissions}{One or more sets of IP permissions. You can't specify a destination security group and a CIDR IP address range in the same set of permissions.}

\item{CidrIp}{Not supported. Use a set of IP permissions to specify the CIDR.}

\item{FromPort}{Not supported. Use a set of IP permissions to specify the port.}

\item{IpProtocol}{Not supported. Use a set of IP permissions to specify the protocol name or number.}

\item{ToPort}{Not supported. Use a set of IP permissions to specify the port.}

\item{SourceSecurityGroupName}{Not supported. Use a set of IP permissions to specify a destination security group.}

\item{SourceSecurityGroupOwnerId}{Not supported. Use a set of IP permissions to specify a destination security group.}
}
\description{
[EC2-VPC only] Removes one or more egress rules from a security group for EC2-VPC. This action doesn't apply to security groups for use in EC2-Classic. To remove a rule, the values that you specify (for example, ports) must match the existing rule's values exactly.
}
\details{
Each rule consists of the protocol and the IPv4 or IPv6 CIDR range or source security group. For the TCP and UDP protocols, you must also specify the destination port or range of ports. For the ICMP protocol, you must also specify the ICMP type and code. If the security group rule has a description, you do not have to specify the description to revoke the rule.

Rule changes are propagated to instances within the security group as quickly as possible. However, a small delay might occur.
}
\examples{

}
