% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{run_instances}
\alias{run_instances}
\title{Launches the specified number of instances using an AMI for which you have permissions}
\usage{
run_instances(MaxCount, MinCount, BlockDeviceMappings = NULL,
  ImageId = NULL, InstanceType = NULL, Ipv6AddressCount = NULL,
  Ipv6Addresses = NULL, KernelId = NULL, KeyName = NULL,
  Monitoring = NULL, Placement = NULL, RamdiskId = NULL,
  SecurityGroupIds = NULL, SecurityGroups = NULL, SubnetId = NULL,
  UserData = NULL, AdditionalInfo = NULL, ClientToken = NULL,
  DisableApiTermination = NULL, DryRun = NULL, EbsOptimized = NULL,
  IamInstanceProfile = NULL, InstanceInitiatedShutdownBehavior = NULL,
  NetworkInterfaces = NULL, PrivateIpAddress = NULL,
  ElasticGpuSpecification = NULL, TagSpecifications = NULL,
  LaunchTemplate = NULL, InstanceMarketOptions = NULL,
  CreditSpecification = NULL, CpuOptions = NULL)
}
\arguments{
\item{MaxCount}{The maximum number of instances to launch. If you specify more instances than Amazon EC2 can launch in the target Availability Zone, Amazon EC2 launches the largest possible number of instances above \code{MinCount}.

Constraints: Between 1 and the maximum number you're allowed for the specified instance type. For more information about the default limits, and how to request an increase, see \href{http://aws.amazon.com/ec2/faqs/#How_many_instances_can_I_run_in_Amazon_EC2}{How many instances can I run in Amazon EC2} in the Amazon EC2 FAQ.}

\item{MinCount}{The minimum number of instances to launch. If you specify a minimum that is more instances than Amazon EC2 can launch in the target Availability Zone, Amazon EC2 launches no instances.

Constraints: Between 1 and the maximum number you're allowed for the specified instance type. For more information about the default limits, and how to request an increase, see \href{http://aws.amazon.com/ec2/faqs/#How_many_instances_can_I_run_in_Amazon_EC2}{How many instances can I run in Amazon EC2} in the Amazon EC2 General FAQ.}

\item{BlockDeviceMappings}{One or more block device mapping entries. You can't specify both a snapshot ID and an encryption value. This is because only blank volumes can be encrypted on creation. If a snapshot is the basis for a volume, it is not blank and its encryption status is used for the volume encryption status.}

\item{ImageId}{The ID of the AMI, which you can get by calling DescribeImages. An AMI is required to launch an instance and must be specified here or in a launch template.}

\item{InstanceType}{The instance type. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/instance-types.html}{Instance Types} in the \emph{Amazon Elastic Compute Cloud User Guide}.

Default: \code{m1.small}}

\item{Ipv6AddressCount}{[EC2-VPC] A number of IPv6 addresses to associate with the primary network interface. Amazon EC2 chooses the IPv6 addresses from the range of your subnet. You cannot specify this option and the option to assign specific IPv6 addresses in the same request. You can specify this option if you've specified a minimum number of instances to launch.}

\item{Ipv6Addresses}{[EC2-VPC] Specify one or more IPv6 addresses from the range of the subnet to associate with the primary network interface. You cannot specify this option and the option to assign a number of IPv6 addresses in the same request. You cannot specify this option if you've specified a minimum number of instances to launch.}

\item{KernelId}{The ID of the kernel.

We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/UserProvidedkernels.html}{PV-GRUB} in the \emph{Amazon Elastic Compute Cloud User Guide}.}

\item{KeyName}{The name of the key pair. You can create a key pair using CreateKeyPair or ImportKeyPair.

If you do not specify a key pair, you can't connect to the instance unless you choose an AMI that is configured to allow users another way to log in.}

\item{Monitoring}{The monitoring for the instance.}

\item{Placement}{The placement for the instance.}

\item{RamdiskId}{The ID of the RAM disk.

We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/UserProvidedkernels.html}{PV-GRUB} in the \emph{Amazon Elastic Compute Cloud User Guide}.}

\item{SecurityGroupIds}{One or more security group IDs. You can create a security group using CreateSecurityGroup.

Default: Amazon EC2 uses the default security group.}

\item{SecurityGroups}{[EC2-Classic, default VPC] One or more security group names. For a nondefault VPC, you must use security group IDs instead.

Default: Amazon EC2 uses the default security group.}

\item{SubnetId}{[EC2-VPC] The ID of the subnet to launch the instance into.}

\item{UserData}{The user data to make available to the instance. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/user-data.html}{Running Commands on Your Linux Instance at Launch} (Linux) and \href{http://docs.aws.amazon.com/AWSEC2/latest/WindowsGuide/ec2-instance-metadata.html#instancedata-add-user-data}{Adding User Data} (Windows). If you are using a command line tool, base64-encoding is performed for you, and you can load the text from a file. Otherwise, you must provide base64-encoded text.}

\item{AdditionalInfo}{Reserved.}

\item{ClientToken}{Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/Run_Instance_Idempotency.html}{Ensuring Idempotency}.

Constraints: Maximum 64 ASCII characters}

\item{DisableApiTermination}{If you set this parameter to \code{true}, you can't terminate the instance using the Amazon EC2 console, CLI, or API; otherwise, you can. To change this attribute to \code{false} after launch, use ModifyInstanceAttribute. Alternatively, if you set \code{InstanceInitiatedShutdownBehavior} to \code{terminate}, you can terminate the instance by running the shutdown command from the instance.

Default: \code{false}}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{EbsOptimized}{Indicates whether the instance is optimized for Amazon EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal Amazon EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS-optimized instance.

Default: \code{false}}

\item{IamInstanceProfile}{The IAM instance profile.}

\item{InstanceInitiatedShutdownBehavior}{Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

Default: \code{stop}}

\item{NetworkInterfaces}{One or more network interfaces.}

\item{PrivateIpAddress}{[EC2-VPC] The primary IPv4 address. You must specify a value from the IPv4 address range of the subnet.

Only one private IP address can be designated as primary. You can't specify this option if you've specified the option to designate a private IP address as the primary IP address in a network interface specification. You cannot specify this option if you're launching more than one instance in the request.}

\item{ElasticGpuSpecification}{An elastic GPU to associate with the instance.}

\item{TagSpecifications}{The tags to apply to the resources during launch. You can only tag instances and volumes on launch. The specified tags are applied to all instances or volumes that are created during launch. To tag a resource after it has been created, see CreateTags.}

\item{LaunchTemplate}{The launch template to use to launch the instances. Any parameters that you specify in RunInstances override the same parameters in the launch template. You can specify either the name or ID of a launch template, but not both.}

\item{InstanceMarketOptions}{The market (purchasing) option for the instances.

For RunInstances, persistent Spot Instance requests are only supported when \strong{InstanceInterruptionBehavior} is set to either \code{hibernate} or \code{stop}.}

\item{CreditSpecification}{The credit option for CPU usage of the instance. Valid values are \code{standard} and \code{unlimited}. To change this attribute after launch, use ModifyInstanceCreditSpecification. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/burstable-performance-instances.html}{Burstable Performance Instances} in the \emph{Amazon Elastic Compute Cloud User Guide}.

Default: \code{standard} (T2 instances) or \code{unlimited} (T3 instances)}

\item{CpuOptions}{The CPU options for the instance. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/instance-optimize-cpu.html}{Optimizing CPU Options} in the \emph{Amazon Elastic Compute Cloud User Guide}.}
}
\description{
Launches the specified number of instances using an AMI for which you have permissions.
}
\details{
You can specify a number of options, or leave the default options. The following rules apply:
\itemize{
\item [EC2-VPC] If you don't specify a subnet ID, we choose a default subnet from your default VPC for you. If you don't have a default VPC, you must specify a subnet ID in the request.
\item [EC2-Classic] If don't specify an Availability Zone, we choose one for you.
\item Some instance types must be launched into a VPC. If you do not have a default VPC, or if you do not specify a subnet ID, the request fails. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/using-vpc.html#vpc-only-instance-types}{Instance Types Available Only in a VPC}.
\item [EC2-VPC] All instances have a network interface with a primary private IPv4 address. If you don't specify this address, we choose one from the IPv4 range of your subnet.
\item Not all instance types support IPv6 addresses. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/instance-types.html}{Instance Types}.
\item If you don't specify a security group ID, we use the default security group. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/using-network-security.html}{Security Groups}.
\item If any of the AMIs have a product code attached for which the user has not subscribed, the request fails.
}

You can create a \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-launch-templates.html}{launch template}, which is a resource that contains the parameters to launch an instance. When you launch an instance using RunInstances, you can specify the launch template instead of specifying the launch parameters.

To ensure faster instance launches, break up large requests into smaller batches. For example, create five separate launch requests for 100 instances each instead of one launch request for 500 instances.

An instance is ready for you to use when it's in the \code{running} state. You can check the state of your instance using DescribeInstances. You can tag instances and EBS volumes during launch, after launch, or both. For more information, see CreateTags and \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/Using_Tags.html}{Tagging Your Amazon EC2 Resources}.

Linux instances have access to the public key of the key pair at boot. You can use this key to provide secure access to the instance. Amazon EC2 public images use this feature to provide secure access without passwords. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-key-pairs.html}{Key Pairs} in the \emph{Amazon Elastic Compute Cloud User Guide}.

For troubleshooting, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/Using_InstanceStraightToTerminated.html}{What To Do If An Instance Immediately Terminates}, and \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/TroubleshootingInstancesConnecting.html}{Troubleshooting Connecting to Your Instance} in the \emph{Amazon Elastic Compute Cloud User Guide}.
}
\examples{
# This example launches an instance using the specified AMI, instance
# type, security group, subnet, block device mapping, and tags.
run_instances(
  BlockDeviceMappings = list(
    list(
      DeviceName = "/dev/sdh",
      Ebs = list(
        VolumeSize = 100L
      )
    )
  ),
  ImageId = "ami-abc12345",
  InstanceType = "t2.micro",
  KeyName = "my-key-pair",
  MaxCount = 1L,
  MinCount = 1L,
  SecurityGroupIds = list(
    "sg-1a2b3c4d"
  ),
  SubnetId = "subnet-6e7f829e",
  TagSpecifications = list(
    list(
      ResourceType = "instance",
      Tags = list(
        list(
          Key = "Purpose",
          Value = "test"
        )
      )
    )
  )
)

}
