% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{run_scheduled_instances}
\alias{run_scheduled_instances}
\title{Launches the specified Scheduled Instances}
\usage{
run_scheduled_instances(LaunchSpecification, ScheduledInstanceId,
  ClientToken = NULL, DryRun = NULL, InstanceCount = NULL)
}
\arguments{
\item{LaunchSpecification}{The launch specification. You must match the instance type, Availability Zone, network, and platform of the schedule that you purchased.}

\item{ScheduledInstanceId}{The Scheduled Instance ID.}

\item{ClientToken}{Unique, case-sensitive identifier that ensures the idempotency of the request. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/Run_Instance_Idempotency.html}{Ensuring Idempotency}.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{InstanceCount}{The number of instances.

Default: 1}
}
\description{
Launches the specified Scheduled Instances.
}
\details{
Before you can launch a Scheduled Instance, you must purchase it and obtain an identifier using PurchaseScheduledInstances.

You must launch a Scheduled Instance during its scheduled time period. You can't stop or reboot a Scheduled Instance, but you can terminate it as needed. If you terminate a Scheduled Instance before the current scheduled time period ends, you can launch it again after a few minutes. For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-scheduled-instances.html}{Scheduled Instances} in the \emph{Amazon Elastic Compute Cloud User Guide}.
}
\examples{
# This example launches the specified Scheduled Instance in a VPC.
run_scheduled_instances(
  InstanceCount = 1L,
  LaunchSpecification = list(
    IamInstanceProfile = list(
      Name = "my-iam-role"
    ),
    ImageId = "ami-12345678",
    InstanceType = "c4.large",
    KeyName = "my-key-pair",
    NetworkInterfaces = list(
      list(
        AssociatePublicIpAddress = TRUE,
        DeviceIndex = 0L,
        Groups = list(
          "sg-12345678"
        ),
        SubnetId = "subnet-12345678"
      )
    )
  ),
  ScheduledInstanceId = "sci-1234-1234-1234-1234-123456789012"
)

# This example launches the specified Scheduled Instance in EC2-Classic.
run_scheduled_instances(
  InstanceCount = 1L,
  LaunchSpecification = list(
    IamInstanceProfile = list(
      Name = "my-iam-role"
    ),
    ImageId = "ami-12345678",
    InstanceType = "c4.large",
    KeyName = "my-key-pair",
    Placement = list(
      AvailabilityZone = "us-west-2b"
    ),
    SecurityGroupIds = list(
      "sg-12345678"
    )
  ),
  ScheduledInstanceId = "sci-1234-1234-1234-1234-123456789012"
)

}
