% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{unassign_ipv_6_addresses}
\alias{unassign_ipv_6_addresses}
\title{Unassigns one or more IPv6 addresses from a network interface}
\usage{
unassign_ipv_6_addresses(Ipv6Addresses, NetworkInterfaceId)
}
\arguments{
\item{Ipv6Addresses}{[required] The IPv6 addresses to unassign from the network interface.}

\item{NetworkInterfaceId}{[required] The ID of the network interface.}
}
\description{
Unassigns one or more IPv6 addresses from a network interface.
}
\section{Accepted Parameters}{
\preformatted{unassign_ipv_6_addresses(
  Ipv6Addresses = list(
    "string"
  ),
  NetworkInterfaceId = "string"
)
}
}

