% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{unmonitor_instances}
\alias{unmonitor_instances}
\title{Disables detailed monitoring for a running instance}
\usage{
unmonitor_instances(InstanceIds, DryRun = NULL)
}
\arguments{
\item{InstanceIds}{[required] One or more instance IDs.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Disables detailed monitoring for a running instance. For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/using-cloudwatch.html}{Monitoring Your Instances and Volumes} in the \emph{Amazon Elastic Compute Cloud User Guide}.
}
\section{Accepted Parameters}{
\preformatted{unmonitor_instances(
  InstanceIds = list(
    "string"
  ),
  DryRun = TRUE|FALSE
)
}
}

