% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{cancel_spot_instance_requests}
\alias{cancel_spot_instance_requests}
\title{Cancels one or more Spot Instance requests}
\usage{
cancel_spot_instance_requests(DryRun = NULL, SpotInstanceRequestIds)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{SpotInstanceRequestIds}{[required] One or more Spot Instance request IDs.}
}
\description{
Cancels one or more Spot Instance requests.
}
\details{
Canceling a Spot Instance request does not terminate running Spot Instances associated with the request.
}
\section{Accepted Parameters}{
\preformatted{cancel_spot_instance_requests(
  DryRun = TRUE|FALSE,
  SpotInstanceRequestIds = list(
    "string"
  )
)
}
}

\examples{
# This example cancels a Spot Instance request.
\donttest{cancel_spot_instance_requests(
  SpotInstanceRequestIds = list(
    "sir-08b93456"
  )
)}

}
