% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{create_placement_group}
\alias{create_placement_group}
\title{Creates a placement group in which to launch instances}
\usage{
create_placement_group(GroupName, Strategy, DryRun = NULL)
}
\arguments{
\item{GroupName}{A name for the placement group. Must be unique within the scope of your account for the region.

Constraints: Up to 255 ASCII characters}

\item{Strategy}{The placement strategy.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Creates a placement group in which to launch instances. The strategy of the placement group determines how the instances are organized within the group.
}
\details{
A \code{cluster} placement group is a logical grouping of instances within a single Availability Zone that benefit from low network latency, high network throughput. A \code{spread} placement group places instances on distinct hardware.

For more information, see \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/placement-groups.html}{Placement Groups} in the \emph{Amazon Elastic Compute Cloud User Guide}.
}
\examples{
# This example creates a placement group with the specified name.
create_placement_group(
  GroupName = "my-cluster",
  Strategy = "cluster"
)

}
