% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{delete_egress_only_internet_gateway}
\alias{delete_egress_only_internet_gateway}
\title{Deletes an egress-only internet gateway}
\usage{
delete_egress_only_internet_gateway(
  DryRun = TRUE|FALSE,
  EgressOnlyInternetGatewayId = "string"
)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{EgressOnlyInternetGatewayId}{[required] The ID of the egress-only internet gateway.}
}
\description{
Deletes an egress-only internet gateway.
}
\examples{

}
