% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{delete_flow_logs}
\alias{delete_flow_logs}
\title{Deletes one or more flow logs}
\usage{
delete_flow_logs(FlowLogIds, DryRun = NULL)
}
\arguments{
\item{FlowLogIds}{One or more flow log IDs.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes one or more flow logs.
}
\examples{

}
