% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{delete_vpc_endpoint_connection_notifications}
\alias{delete_vpc_endpoint_connection_notifications}
\title{Deletes one or more VPC endpoint connection notifications}
\usage{
delete_vpc_endpoint_connection_notifications(ConnectionNotificationIds,
  DryRun = NULL)
}
\arguments{
\item{ConnectionNotificationIds}{One or more notification IDs.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Deletes one or more VPC endpoint connection notifications.
}
\examples{

}
