% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{describe_transit_gateway_vpc_attachments}
\alias{describe_transit_gateway_vpc_attachments}
\title{Describes one or more VPC attachments}
\usage{

  describe_transit_gateway_vpc_attachments(TransitGatewayAttachmentIds = NULL,
  Filters = NULL, MaxResults = NULL, NextToken = NULL,
  DryRun = NULL)
}
\arguments{
\item{TransitGatewayAttachmentIds}{The IDs of the attachments.}

\item{Filters}{One or more filters. The possible values are:
\itemize{
\item \code{state} - The state of the attachment (\code{available} | \code{deleted} | \code{deleting} | \code{failed} | \code{modifying} | \code{pendingAcceptance} | \code{pending} | \code{rollingBack} | \code{rejected} | \code{rejecting}).
\item \code{transit-gateway-attachment-id} - The ID of the attachment.
\item \code{transit-gateway-id} - The ID of the transit gateway.
\item \code{vpc-id} - The ID of the VPC.
}}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned \code{nextToken} value.}

\item{NextToken}{The token for the next page of results.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Describes one or more VPC attachments. By default, all VPC attachments are described. Alternatively, you can filter the results.
}
\section{Accepted Parameters}{
\preformatted{describe_transit_gateway_vpc_attachments(
  TransitGatewayAttachmentIds = list(
    "string"
  ),
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxResults = 123,
  NextToken = "string",
  DryRun = TRUE|FALSE
)
}
}

