% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{enable_vgw_route_propagation}
\alias{enable_vgw_route_propagation}
\title{Enables a virtual private gateway (VGW) to propagate routes to the specified route table of a VPC}
\usage{
enable_vgw_route_propagation(GatewayId, RouteTableId)
}
\arguments{
\item{GatewayId}{The ID of the virtual private gateway.}

\item{RouteTableId}{The ID of the route table.}
}
\description{
Enables a virtual private gateway (VGW) to propagate routes to the specified route table of a VPC.
}
\examples{
# This example enables the specified virtual private gateway to propagate
# static routes to the specified route table.
enable_vgw_route_propagation(
  GatewayId = "vgw-9a4cacf3",
  RouteTableId = "rtb-22574640"
)

}
