% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{get_launch_template_data}
\alias{get_launch_template_data}
\title{Retrieves the configuration data of the specified instance}
\usage{
get_launch_template_data(InstanceId, DryRun = NULL)
}
\arguments{
\item{InstanceId}{The ID of the instance.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Retrieves the configuration data of the specified instance. You can use this data to create a launch template.
}
\examples{
# This example gets the launch template data for the specified instance.
get_launch_template_data(
  InstanceId = "0123d646e8048babc"
)

}
