% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{modify_vpc_peering_connection_options}
\alias{modify_vpc_peering_connection_options}
\title{Modifies the VPC peering connection options on one side of a VPC peering connection}
\usage{
modify_vpc_peering_connection_options(VpcPeeringConnectionId,
  AccepterPeeringConnectionOptions = NULL, DryRun = NULL,
  RequesterPeeringConnectionOptions = NULL)
}
\arguments{
\item{VpcPeeringConnectionId}{The ID of the VPC peering connection.}

\item{AccepterPeeringConnectionOptions}{The VPC peering connection options for the accepter VPC.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}

\item{RequesterPeeringConnectionOptions}{The VPC peering connection options for the requester VPC.}
}
\description{
Modifies the VPC peering connection options on one side of a VPC peering connection. You can do the following:
}
\details{
\itemize{
\item Enable/disable communication over the peering connection between an EC2-Classic instance that's linked to your VPC (using ClassicLink) and instances in the peer VPC.
\item Enable/disable communication over the peering connection between instances in your VPC and an EC2-Classic instance that's linked to the peer VPC.
\item Enable/disable the ability to resolve public DNS hostnames to private IP addresses when queried from instances in the peer VPC.
}

If the peered VPCs are in different accounts, each owner must initiate a separate request to modify the peering connection options, depending on whether their VPC was the requester or accepter for the VPC peering connection. If the peered VPCs are in the same account, you can modify the requester and accepter options in the same request. To confirm which VPC is the accepter and requester for a VPC peering connection, use the DescribeVpcPeeringConnections command.
}
\examples{

}
