% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{reject_vpc_endpoint_connections}
\alias{reject_vpc_endpoint_connections}
\title{Rejects one or more VPC endpoint connection requests to your VPC endpoint service}
\usage{
reject_vpc_endpoint_connections(ServiceId, VpcEndpointIds, DryRun = NULL)
}
\arguments{
\item{ServiceId}{The ID of the service.}

\item{VpcEndpointIds}{The IDs of one or more VPC endpoints.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Rejects one or more VPC endpoint connection requests to your VPC endpoint service.
}
\examples{

}
