% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ec2_operations.R
\name{search_transit_gateway_routes}
\alias{search_transit_gateway_routes}
\title{Searches for routes in the specified transit gateway route table}
\usage{
search_transit_gateway_routes(
  TransitGatewayRouteTableId = "string",
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxResults = 123,
  DryRun = TRUE|FALSE
)
}
\arguments{
\item{TransitGatewayRouteTableId}{[required] The ID of the transit gateway route table.}

\item{Filters}{[required] One or more filters. The possible values are:
\itemize{
\item \code{transit-gateway-route-destination-cidr-block} - The CIDR range.
\item \code{transit-gateway-route-state} - The state of the route (\code{active} | \code{blackhole}).
\item \code{transit-gateway-route-transit-gateway-attachment-id} - The ID of the attachment.
\item \code{transit-gateway-route-type} - The route type (\code{static} | \code{propagated}).
\item \code{transit-gateway-route-vpn-connection-id} - The ID of the VPN connection.
}}

\item{MaxResults}{The maximum number of routes to return.}

\item{DryRun}{Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is \code{DryRunOperation}. Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Searches for routes in the specified transit gateway route table.
}
\examples{

}
